refactor(http): split transport/http.rs into session, stateless_dispatch, and handlers
Phase 3 of #1256. transport/http.rs was 4869 lines after phase 1 (#1318)
moved its inline tests to a sibling tests.rs. This splits the remaining
production code into three child modules, following the pattern already
used for router.rs in phase 2 (router/task_ops.rs, router/notify.rs):

- session.rs: Session, SessionRegistry, SessionConfig, SessionHandle,
  SessionInfo, and the internal SessionServiceSource/PendingRequest types.
- stateless_dispatch.rs: the 2026-07-28 stateless-protocol-specific
  machinery (ModernSubscriptionRegistry and friends, the modern
  subscriptions/listen handler, per-request _meta extraction, the SSE
  upgrade for a sessionless POST). Gated as a whole module
  (#[cfg(feature = "stateless")] on the mod declaration) instead of
  per item, since every item in it needs the feature; this also removes
  the risk of an item's own #[cfg] going stale after a future edit.
- handlers.rs: the axum route handlers (handle_post, handle_get,
  handle_delete, handle_health, the legacy subscriptions/listen SSE
  path) plus the request-validation and response-building helpers they
  share. handle_post stays one function; it already carries multiple
  protocol lifecycles and dispatch paths inline (#1242 touched five
  separate places in it), and untangling that is out of scope for a
  structural move.

http.rs keeps the HttpTransport struct and its entire impl (construction,
every builder method, into_router*, serve*), AppState, ServiceSource, and
OAuthConfig -- the same role router.rs kept for McpRouter's own dispatch
during phase 2.

No public API change: SessionConfig, SessionHandle, SessionInfo, and
DEFAULT_SESSION_TTL are re-exported from http.rs at their existing paths.
Internal items moved to pub(super), narrowed further where a moved item
turned out to have no caller outside its own module (e.g. ModernSubscription
and ModernSubscriptionGuard stay fully private to stateless_dispatch.rs).

Test count unchanged: 1029 lib tests before and after (98 in
transport::http::tests::, also unchanged).
