+++ crates/tower-mcp/src/transport/http.rs
@@ -292,19 +292,6 @@ use crate::transport::subscriptions::{
 use crate::{ProtocolSupport, ProtocolSupportError};
 use tower::util::BoxCloneService;
 
-/// SEP-2575 per-request `_meta` extraction. Pulls `StatelessRequestMeta` from
-/// the parsed request params and inserts it into the per-request `Extensions`
-/// so handlers can read it via `ctx.per_request_meta()`. No-op if the request
-/// has no `_meta`, params aren't an object, or the meta can't deserialize.
-#[cfg(feature = "stateless")]
-fn stash_per_request_meta(req: &JsonRpcRequest, ext: &mut crate::router::Extensions) {
-    if let Some(params) = req.params.as_ref()
-        && let Some(meta) = crate::stateless::StatelessRequestMeta::from_params(params)
-    {
-        ext.insert(meta);
-    }
-}
-
 /// Header name for MCP session ID
 pub const MCP_SESSION_ID_HEADER: &str = "mcp-session-id";
 
@@ -338,976 +325,6 @@ const SSE_MESSAGE_EVENT: &str = "message";
 
 /// Header name for Last-Event-ID (for SSE stream resumption per SEP-1699)
 const LAST_EVENT_ID_HEADER: &str = "last-event-id";
-
-/// Pending request waiting for a response from the client
-struct PendingRequest {
-    response_tx: oneshot::Sender<Result<serde_json::Value>>,
-}
-
-type AssociatedCall = Pin<Box<dyn Future<Output = Result<JsonRpcResponse>> + Send + 'static>>;
-
-/// Session state for HTTP transport
-/// How a session produces its MCP service for request processing.
-enum SessionServiceSource {
-    /// Session was created from an McpRouter with a factory for middleware wrapping.
-    Router {
-        router: McpRouter,
-        factory: ServiceFactory,
-    },
-    /// Session was created from a pre-built boxed service (e.g., McpProxy).
-    /// Wrapped in Mutex because BoxCloneService is Send but not Sync,
-    /// and Session must be Sync for Arc<Session> to be Send.
-    Boxed(std::sync::Mutex<McpBoxService>),
-}
-
-struct Session {
-    /// Session ID
-    id: String,
-    /// Source for creating the MCP service
-    service_source: SessionServiceSource,
-    /// Broadcast channel for SSE notifications and outgoing requests
-    notifications_tx: broadcast::Sender<String>,
-    /// When this session was created
-    created_at: Instant,
-    /// Last time this session was accessed
-    last_accessed: RwLock<Instant>,
-    /// Pending outgoing requests waiting for responses
-    pending_requests: Mutex<HashMap<RequestId, PendingRequest>>,
-    /// Session-wide allocator for request-scoped server-to-client request IDs.
-    ///
-    /// Each originating POST owns a separate channel, but IDs must remain
-    /// unique across concurrent POSTs in the same session.
-    request_id_allocator: Option<Arc<AtomicI64>>,
-    /// Negotiated protocol version (set after initialize)
-    protocol_version: RwLock<String>,
-    /// Client implementation info advertised in the `initialize` request.
-    ///
-    /// Populated by `handle_post` after a successful initialize response,
-    /// and restored from a [`SessionRecord`](crate::session_store::SessionRecord)
-    /// when a session is rebuilt from the persistent store. `None` until the
-    /// first initialize completes.
-    client_info: RwLock<Option<Implementation>>,
-    /// Client capabilities advertised in the `initialize` request.
-    ///
-    /// Populated by `handle_post` after a successful initialize response,
-    /// and restored from a [`SessionRecord`](crate::session_store::SessionRecord)
-    /// when a session is rebuilt from the persistent store. `None` until the
-    /// first initialize completes.
-    client_capabilities: RwLock<Option<ClientCapabilities>>,
-    /// Counter for SSE event IDs (for stream resumption per SEP-1699)
-    event_counter: AtomicU64,
-    /// Pluggable store for SSE events (enables cross-instance replay)
-    event_store: Arc<dyn crate::event_store::EventStore>,
-    /// Whether `notifications/initialized` has been received from the client.
-    ///
-    /// Per the MCP 2025-11-25 spec, clients MUST send this notification after
-    /// receiving the `initialize` response and before sending any other requests.
-    /// Checked by `handle_post` when `strict_initialization` is enabled on
-    /// [`SessionConfig`]. Pre-initialized sessions (optional_sessions path) and
-    /// restored sessions start with this set to `true`.
-    initialized_notification_received: std::sync::atomic::AtomicBool,
-}
-
-impl Session {
-    fn new(
-        router: McpRouter,
-        sampling_enabled: bool,
-        service_factory: ServiceFactory,
-        event_store: Arc<dyn crate::event_store::EventStore>,
-    ) -> Self {
-        let (notifications_tx, _) = broadcast::channel(100);
-
-        // Set up notification forwarding: mpsc -> broadcast
-        // The router sends notifications (progress, log, resource updates) to
-        // an mpsc channel. We bridge these to the session's broadcast channel
-        // so they reach connected SSE clients.
-        let (notif_sender, mut notif_receiver) = notification_channel(256);
-        let router = router.with_notification_sender(notif_sender);
-
-        let broadcast_tx = notifications_tx.clone();
-        tokio::spawn(async move {
-            while let Some(notification) = notif_receiver.recv().await {
-                if let Some(json) = crate::transport::stdio::serialize_notification(&notification) {
-                    // Best effort: if no subscribers, the message is dropped
-                    let _ = broadcast_tx.send(json);
-                }
-            }
-        });
-
-        let request_id_allocator = if sampling_enabled {
-            Some(Arc::new(AtomicI64::new(1)))
-        } else {
-            None
-        };
-
-        let now = Instant::now();
-        Self {
-            id: uuid::Uuid::new_v4().to_string(),
-            service_source: SessionServiceSource::Router {
-                router,
-                factory: service_factory,
-            },
-            notifications_tx,
-            created_at: now,
-            last_accessed: RwLock::new(now),
-            pending_requests: Mutex::new(HashMap::new()),
-            request_id_allocator,
-            protocol_version: RwLock::new(LATEST_PROTOCOL_VERSION.to_string()),
-            client_info: RwLock::new(None),
-            client_capabilities: RwLock::new(None),
-            event_counter: AtomicU64::new(0),
-            event_store,
-            initialized_notification_received: std::sync::atomic::AtomicBool::new(false),
-        }
-    }
-
-    /// Create a session from a pre-built boxed service.
-    ///
-    /// This is used when the transport is created via [`HttpTransport::from_service()`].
-    /// Notification bridging and sampling setup are skipped — the caller is
-    /// responsible for configuring these on the service before passing it in.
-    fn from_service(
-        service: McpBoxService,
-        event_store: Arc<dyn crate::event_store::EventStore>,
-    ) -> Self {
-        let (notifications_tx, _) = broadcast::channel(100);
-
-        let now = Instant::now();
-        Self {
-            id: uuid::Uuid::new_v4().to_string(),
-            service_source: SessionServiceSource::Boxed(std::sync::Mutex::new(service)),
-            notifications_tx,
-            created_at: now,
-            last_accessed: RwLock::new(now),
-            pending_requests: Mutex::new(HashMap::new()),
-            request_id_allocator: None,
-            protocol_version: RwLock::new(LATEST_PROTOCOL_VERSION.to_string()),
-            client_info: RwLock::new(None),
-            client_capabilities: RwLock::new(None),
-            event_counter: AtomicU64::new(0),
-            event_store,
-            initialized_notification_received: std::sync::atomic::AtomicBool::new(false),
-        }
-    }
-
-    /// Rebuild a session from a [`SessionRecord`] so a request for an
-    /// unknown session ID can be served transparently.
-    ///
-    /// The router is pre-marked initialized and the protocol version is
-    /// restored from the record. Runtime state (broadcast channels,
-    /// pending-request table) is freshly allocated — in-flight state from
-    /// before the rebuild is not recovered. The `event_counter` is left at
-    /// zero; the [`SessionRegistry`] seeds it from the event store so
-    /// future event IDs don't collide with buffered ones.
-    fn restored(
-        record: &crate::session_store::SessionRecord,
-        router: McpRouter,
-        sampling_enabled: bool,
-        service_factory: ServiceFactory,
-        event_store: Arc<dyn crate::event_store::EventStore>,
-    ) -> Self {
-        // Skip the Initializing intermediate state — this session was
-        // already initialized on the original instance.
-        router.session().mark_preinitialized();
-
-        let (notifications_tx, _) = broadcast::channel(100);
-        let (notif_sender, mut notif_receiver) = notification_channel(256);
-        let router = router.with_notification_sender(notif_sender);
-
-        let broadcast_tx = notifications_tx.clone();
-        tokio::spawn(async move {
-            while let Some(notification) = notif_receiver.recv().await {
-                if let Some(json) = crate::transport::stdio::serialize_notification(&notification) {
-                    let _ = broadcast_tx.send(json);
-                }
-            }
-        });
-
-        let request_id_allocator = if sampling_enabled {
-            Some(Arc::new(AtomicI64::new(1)))
-        } else {
-            None
-        };
-
-        let now = Instant::now();
-        Self {
-            id: record.id.clone(),
-            service_source: SessionServiceSource::Router {
-                router,
-                factory: service_factory,
-            },
-            notifications_tx,
-            created_at: now,
-            last_accessed: RwLock::new(now),
-            pending_requests: Mutex::new(HashMap::new()),
-            request_id_allocator,
-            protocol_version: RwLock::new(record.protocol_version.clone()),
-            client_info: RwLock::new(record.client_info.clone()),
-            client_capabilities: RwLock::new(record.client_capabilities.clone()),
-            event_counter: AtomicU64::new(0),
-            event_store,
-            // Restored sessions already completed the handshake on a previous
-            // instance; treat `notifications/initialized` as already received.
-            initialized_notification_received: std::sync::atomic::AtomicBool::new(true),
-        }
-    }
-
-    /// Rebuild a session from a [`SessionRecord`] for transports built
-    /// with [`HttpTransport::from_service`]. The service's internal state
-    /// (if any) is not restored — the caller is responsible for anything
-    /// beyond the metadata in the record.
-    fn from_service_restored(
-        service: McpBoxService,
-        record: &crate::session_store::SessionRecord,
-        event_store: Arc<dyn crate::event_store::EventStore>,
-    ) -> Self {
-        let (notifications_tx, _) = broadcast::channel(100);
-        let now = Instant::now();
-        Self {
-            id: record.id.clone(),
-            service_source: SessionServiceSource::Boxed(std::sync::Mutex::new(service)),
-            notifications_tx,
-            created_at: now,
-            last_accessed: RwLock::new(now),
-            pending_requests: Mutex::new(HashMap::new()),
-            request_id_allocator: None,
-            protocol_version: RwLock::new(record.protocol_version.clone()),
-            client_info: RwLock::new(record.client_info.clone()),
-            client_capabilities: RwLock::new(record.client_capabilities.clone()),
-            event_counter: AtomicU64::new(0),
-            event_store,
-            // Restored sessions already completed the handshake on a previous
-            // instance; treat `notifications/initialized` as already received.
-            initialized_notification_received: std::sync::atomic::AtomicBool::new(true),
-        }
-    }
-
-    /// Create a middleware-wrapped service from this session's service source.
-    fn make_service(&self) -> McpBoxService {
-        match &self.service_source {
-            SessionServiceSource::Router { router, factory } => (factory)(router.clone()),
-            SessionServiceSource::Boxed(mutex) => mutex.lock().unwrap().clone(),
-        }
-    }
-
-    /// Handle a client notification (fire-and-forget, no response).
-    ///
-    /// For router-based sessions, delegates to the router's notification handler.
-    /// For service-based sessions, notifications are logged but not processed
-    /// (the service should handle its own notification needs).
-    fn handle_notification(&self, notification: McpNotification) {
-        match &self.service_source {
-            SessionServiceSource::Router { router, .. } => {
-                router.handle_notification(notification);
-            }
-            SessionServiceSource::Boxed(_) => {
-                tracing::debug!(
-                    notification = ?notification,
-                    "Notification received on service-based session (not forwarded)"
-                );
-            }
-        }
-    }
-
-    /// Get the next SSE event ID for this session.
-    ///
-    /// Event IDs are monotonically increasing per session, enabling
-    /// stream resumption via the Last-Event-ID header (SEP-1699).
-    fn next_event_id(&self) -> u64 {
-        self.event_counter.fetch_add(1, Ordering::SeqCst)
-    }
-
-    /// Buffer an event for potential replay (SEP-1699).
-    ///
-    /// Delegates to the configured [`EventStore`](crate::event_store::EventStore).
-    /// Store errors are logged but non-fatal — the transport continues
-    /// serving the client even if the external event buffer is unavailable,
-    /// since the event has already been sent on the live SSE stream.
-    async fn buffer_event(&self, id: u64, data: String) {
-        let record = crate::event_store::EventRecord::new(id, data);
-        if let Err(e) = self.event_store.append(&self.id, record).await {
-            tracing::warn!(session_id = %self.id, event_id = id, error = %e, "Failed to append event to event store");
-        }
-    }
-
-    /// Get buffered events after the given event ID.
-    ///
-    /// Returns events with IDs greater than `after_id`, in order. Used for
-    /// stream resumption when a client reconnects with the `Last-Event-ID`
-    /// header. Store errors produce an empty replay list and are logged.
-    async fn get_events_after(&self, after_id: u64) -> Vec<crate::event_store::EventRecord> {
-        match self.event_store.replay_after(&self.id, after_id).await {
-            Ok(events) => events,
-            Err(e) => {
-                tracing::warn!(session_id = %self.id, error = %e, "Failed to replay events from event store");
-                Vec::new()
-            }
-        }
-    }
-
-    /// Update the last accessed time
-    async fn touch(&self) {
-        *self.last_accessed.write().await = Instant::now();
-    }
-
-    /// Check if the session has expired
-    async fn is_expired(&self, ttl: Duration) -> bool {
-        self.last_accessed.read().await.elapsed() > ttl
-    }
-
-    /// Store a pending request
-    async fn add_pending_request(
-        &self,
-        id: RequestId,
-        response_tx: oneshot::Sender<Result<serde_json::Value>>,
-    ) {
-        let mut pending = self.pending_requests.lock().await;
-        pending.insert(id, PendingRequest { response_tx });
-    }
-
-    /// Complete a pending request with a response
-    async fn complete_pending_request(
-        &self,
-        id: &RequestId,
-        result: Result<serde_json::Value>,
-    ) -> bool {
-        let pending = {
-            let mut pending_requests = self.pending_requests.lock().await;
-            pending_requests.remove(id)
-        };
-
-        match pending {
-            Some(pending) => {
-                // Send result to waiter (ignore if they've dropped the receiver)
-                let _ = pending.response_tx.send(result);
-                true
-            }
-            None => false,
-        }
-    }
-
-    /// Fail request-scoped client requests whose originating POST is gone.
-    async fn fail_pending_requests(&self, ids: &[RequestId], message: &str) {
-        let removed = {
-            let mut pending = self.pending_requests.lock().await;
-            ids.iter()
-                .filter_map(|id| pending.remove(id))
-                .collect::<Vec<_>>()
-        };
-
-        for pending in removed {
-            let _ = pending
-                .response_tx
-                .send(Err(Error::Transport(message.to_string())));
-        }
-    }
-}
-
-/// Default session TTL (30 minutes)
-pub const DEFAULT_SESSION_TTL: Duration = Duration::from_secs(30 * 60);
-
-/// Default cleanup interval (1 minute)
-const DEFAULT_CLEANUP_INTERVAL: Duration = Duration::from_secs(60);
-
-/// Configuration for session management
-#[derive(Debug, Clone)]
-pub struct SessionConfig {
-    /// Time-to-live for inactive sessions
-    pub ttl: Duration,
-    /// Maximum number of sessions (None = unlimited)
-    pub max_sessions: Option<usize>,
-    /// How often to run the cleanup task
-    pub cleanup_interval: Duration,
-    /// Whether to enforce that clients send `notifications/initialized` before
-    /// making any non-initialize requests, per the MCP 2025-11-25 spec.
-    ///
-    /// When `true` (the default), the transport returns a JSON-RPC
-    /// `InvalidRequest` error (-32600) to any request received before
-    /// `notifications/initialized` on a 2025-11-25 session-based connection.
-    ///
-    /// Set to `false` to restore the previous lenient behavior, e.g. in
-    /// dev/test scenarios where the full MCP handshake is inconvenient.
-    pub strict_initialization: bool,
-}
-
-impl Default for SessionConfig {
-    fn default() -> Self {
-        Self {
-            ttl: DEFAULT_SESSION_TTL,
-            max_sessions: None,
-            cleanup_interval: DEFAULT_CLEANUP_INTERVAL,
-            strict_initialization: true,
-        }
-    }
-}
-
-impl SessionConfig {
-    /// Create a new session config with the given TTL
-    pub fn with_ttl(ttl: Duration) -> Self {
-        Self {
-            ttl,
-            ..Default::default()
-        }
-    }
-
-    /// Set the maximum number of sessions
-    pub fn max_sessions(mut self, max: usize) -> Self {
-        self.max_sessions = Some(max);
-        self
-    }
-
-    /// Set the cleanup interval
-    pub fn cleanup_interval(mut self, interval: Duration) -> Self {
-        self.cleanup_interval = interval;
-        self
-    }
-
-    /// Enable or disable strict initialization enforcement.
-    ///
-    /// When enabled (default), the transport enforces that clients send
-    /// `notifications/initialized` before any other requests on a
-    /// 2025-11-25 session-based connection, per the MCP spec. Requests
-    /// that arrive before this notification receive a JSON-RPC
-    /// `InvalidRequest` error (-32600).
-    ///
-    /// Disable this for dev/test scenarios where the full MCP handshake
-    /// is inconvenient.
-    pub fn strict_initialization(mut self, enabled: bool) -> Self {
-        self.strict_initialization = enabled;
-        self
-    }
-}
-
-/// Registry coordinating live session runtime state with a pluggable
-/// persistent [`SessionStore`](crate::session_store::SessionStore).
-///
-/// - Runtime state (broadcast channels, pending requests, live services) is
-///   kept in the in-process `sessions` map and cannot be serialized.
-/// - Persistent metadata (IDs, timestamps, protocol version) is mirrored into
-///   the caller-supplied [`SessionStore`]. The default
-///   [`MemorySessionStore`](crate::session_store::MemorySessionStore) keeps
-///   metadata in-process (same behavior as before this trait existed).
-struct SessionRegistry {
-    sessions: RwLock<HashMap<String, Arc<Session>>>,
-    config: SessionConfig,
-    sampling_enabled: bool,
-    persistent: Arc<dyn crate::session_store::SessionStore>,
-    events: Arc<dyn crate::event_store::EventStore>,
-    /// Source for rebuilding services when restoring a session.
-    service_source: ServiceSource,
-    /// If `true`, a request for an unknown session ID whose record is not
-    /// in the persistent store spins up a new session with synthetic
-    /// client info instead of returning 404 (see anubis-mcp #125 for the
-    /// precedent).
-    auto_reinit: bool,
-}
-
-impl SessionRegistry {
-    fn new(
-        config: SessionConfig,
-        sampling_enabled: bool,
-        persistent: Arc<dyn crate::session_store::SessionStore>,
-        events: Arc<dyn crate::event_store::EventStore>,
-        service_source: ServiceSource,
-        auto_reinit: bool,
-    ) -> Self {
-        Self {
-            sessions: RwLock::new(HashMap::new()),
-            config,
-            sampling_enabled,
-            persistent,
-            events,
-            service_source,
-            auto_reinit,
-        }
-    }
-
-    /// Build a SessionRecord reflecting the given live Session.
-    async fn record_for(&self, session: &Session) -> crate::session_store::SessionRecord {
-        let protocol_version = session.protocol_version.read().await.clone();
-        let last_accessed = session.last_accessed.read().await;
-        let mut record = crate::session_store::SessionRecord::new(
-            session.id.clone(),
-            protocol_version,
-            self.config.ttl,
-        );
-        // Populate the client identity / capabilities advertised at
-        // initialize time so persisted records faithfully describe the
-        // session. These remain `None` until a successful initialize.
-        record.client_info = session.client_info.read().await.clone();
-        record.client_capabilities = session.client_capabilities.read().await.clone();
-        // Convert from monotonic Instant to SystemTime approximation.
-        let now = std::time::SystemTime::now();
-        let created_ago = session.created_at.elapsed();
-        let last_accessed_ago = last_accessed.elapsed();
-        record.created_at = now.checked_sub(created_ago).unwrap_or(now);
-        record.last_accessed = now.checked_sub(last_accessed_ago).unwrap_or(now);
-        record.expires_at = record.last_accessed + self.config.ttl;
-        record
-    }
-
-    /// Persist metadata for a newly created session, logging on failure.
-    ///
-    /// Persistence errors are intentionally non-fatal: the live runtime
-    /// session is already registered locally, so the transport can continue
-    /// serving requests even if the external store is briefly unavailable.
-    async fn persist_new(&self, session: &Session) {
-        let record = self.record_for(session).await;
-        if let Err(e) = self.persistent.create(&mut record.clone()).await {
-            tracing::warn!(session_id = %session.id, error = %e, "Failed to persist session record");
-        }
-    }
-
-    /// Persist an update to an existing session's record (upsert).
-    ///
-    /// Called after the session's state changes in a way that should be
-    /// reflected in the persistent store -- notably after a successful
-    /// `initialize` so the stored record carries the client's advertised
-    /// `client_info` and `capabilities` (rather than the defaults captured
-    /// at create time). Failures are logged but non-fatal.
-    async fn save_record(&self, session: &Session) {
-        let record = self.record_for(session).await;
-        if let Err(e) = self.persistent.save(&record).await {
-            tracing::warn!(session_id = %session.id, error = %e, "Failed to save session record");
-        }
-    }
-
-    /// Create a session for an incoming `initialize` request.
-    ///
-    /// Only reached from the `is_init` branch of the POST handler, so the
-    /// handshake is recorded here, before the request is dispatched. A client
-    /// whose `initialized` notification overtakes that dispatch (#458) then
-    /// still completes the handshake, while one that never sent `initialize`
-    /// has no session for the notification to open.
-    async fn create(
-        &self,
-        router: McpRouter,
-        service_factory: ServiceFactory,
-    ) -> Option<Arc<Session>> {
-        router.session().mark_handshake_started();
-
-        let session = {
-            let mut sessions = self.sessions.write().await;
-
-            // Check max sessions limit
-            if let Some(max) = self.config.max_sessions
-                && sessions.len() >= max
-            {
-                tracing::warn!(
-                    max_sessions = max,
-                    current = sessions.len(),
-                    "Session limit reached, rejecting new session"
-                );
-                return None;
-            }
-
-            let session = Arc::new(Session::new(
-                router,
-                self.sampling_enabled,
-                service_factory,
-                self.events.clone(),
-            ));
-            sessions.insert(session.id.clone(), session.clone());
-            tracing::debug!(session_id = %session.id, sampling = self.sampling_enabled, "Created new session");
-            session
-        };
-        self.persist_new(&session).await;
-        Some(session)
-    }
-
-    async fn create_from_service(&self, service: McpBoxService) -> Option<Arc<Session>> {
-        let session = {
-            let mut sessions = self.sessions.write().await;
-
-            if let Some(max) = self.config.max_sessions
-                && sessions.len() >= max
-            {
-                tracing::warn!(
-                    max_sessions = max,
-                    current = sessions.len(),
-                    "Session limit reached, rejecting new session"
-                );
-                return None;
-            }
-
-            let session = Arc::new(Session::from_service(service, self.events.clone()));
-            sessions.insert(session.id.clone(), session.clone());
-            tracing::debug!(session_id = %session.id, "Created new session from service");
-            session
-        };
-        self.persist_new(&session).await;
-        Some(session)
-    }
-
-    /// Create a new session with its router already marked as initialized.
-    ///
-    /// Used by the optional-sessions feature to serve requests from clients
-    /// that skip the initialize handshake.
-    async fn create_initialized(
-        &self,
-        router: McpRouter,
-        service_factory: ServiceFactory,
-    ) -> Option<Arc<Session>> {
-        // Pre-initialize the router's session state so it won't reject requests
-        router.session().mark_preinitialized();
-
-        let session = {
-            let mut sessions = self.sessions.write().await;
-
-            if let Some(max) = self.config.max_sessions
-                && sessions.len() >= max
-            {
-                return None;
-            }
-
-            let session = Arc::new(Session::new(
-                router,
-                self.sampling_enabled,
-                service_factory,
-                self.events.clone(),
-            ));
-            // Pre-initialized sessions bypass the full MCP handshake (they
-            // exist for clients that don't track session IDs). Mark the
-            // notification as already received so strict_initialization checks
-            // don't reject their requests.
-            session
-                .initialized_notification_received
-                .store(true, Ordering::Release);
-            sessions.insert(session.id.clone(), session.clone());
-            tracing::debug!(session_id = %session.id, "Created pre-initialized session (optional_sessions)");
-            session
-        };
-        self.persist_new(&session).await;
-        Some(session)
-    }
-
-    /// Create a pre-initialized session from a boxed service.
-    async fn create_initialized_from_service(
-        &self,
-        service: McpBoxService,
-    ) -> Option<Arc<Session>> {
-        let session = {
-            let mut sessions = self.sessions.write().await;
-
-            if let Some(max) = self.config.max_sessions
-                && sessions.len() >= max
-            {
-                return None;
-            }
-
-            let session = Arc::new(Session::from_service(service, self.events.clone()));
-            // Pre-initialized sessions bypass the full MCP handshake; mark the
-            // notification as already received.
-            session
-                .initialized_notification_received
-                .store(true, Ordering::Release);
-            sessions.insert(session.id.clone(), session.clone());
-            tracing::debug!(session_id = %session.id, "Created pre-initialized session from service (optional_sessions)");
-            session
-        };
-        self.persist_new(&session).await;
-        Some(session)
-    }
-
-    async fn get(&self, id: &str) -> Option<Arc<Session>> {
-        // Fast path: the session is live in this process.
-        {
-            let sessions = self.sessions.read().await;
-            if let Some(s) = sessions.get(id).cloned() {
-                s.touch().await;
-                return Some(s);
-            }
-        }
-
-        // Slow path #1: the session is unknown locally but the persistent
-        // store has a record — rebuild it.
-        match self.persistent.load(id).await {
-            Ok(Some(record)) => {
-                tracing::info!(session_id = %id, "Restoring session from persistent store");
-                if let Some(session) = self.restore_from_record(record).await {
-                    return Some(session);
-                }
-            }
-            Ok(None) => {}
-            Err(e) => {
-                tracing::warn!(session_id = %id, error = %e, "Failed to load session record");
-            }
-        }
-
-        // Slow path #2 (opt-in): auto-reinitialize with synthetic client
-        // info so the client can continue without a re-handshake. Useful
-        // for single-instance restarts where no external store is
-        // configured; loses original client identity.
-        if self.auto_reinit {
-            tracing::info!(session_id = %id, "Auto-reinitializing unknown session");
-            return self.auto_reinitialize(id).await;
-        }
-
-        None
-    }
-
-    /// Restore a live [`Session`] from a persisted [`SessionRecord`].
-    ///
-    /// The caller must ensure the record's ID is not already live locally;
-    /// on success the session is inserted into the local registry, the
-    /// event counter is seeded so new event IDs don't collide with
-    /// buffered ones, and the record's `last_accessed` is refreshed and
-    /// saved back to the store.
-    async fn restore_from_record(
-        &self,
-        record: crate::session_store::SessionRecord,
-    ) -> Option<Arc<Session>> {
-        let session = {
-            let mut sessions = self.sessions.write().await;
-
-            if let Some(max) = self.config.max_sessions
-                && sessions.len() >= max
-            {
-                tracing::warn!(
-                    max_sessions = max,
-                    "Session limit reached, cannot restore session"
-                );
-                return None;
-            }
-
-            // Guard against a concurrent create that beat us here.
-            if let Some(existing) = sessions.get(&record.id).cloned() {
-                existing.touch().await;
-                return Some(existing);
-            }
-
-            let session: Arc<Session> = match &self.service_source {
-                ServiceSource::Router { router, factory } => Arc::new(Session::restored(
-                    &record,
-                    router.with_fresh_session(),
-                    self.sampling_enabled,
-                    factory.clone(),
-                    self.events.clone(),
-                )),
-                ServiceSource::Service(svc) => {
-                    let service = svc.lock().unwrap().clone();
-                    Arc::new(Session::from_service_restored(
-                        service,
-                        &record,
-                        self.events.clone(),
-                    ))
-                }
-            };
-
-            sessions.insert(record.id.clone(), session.clone());
-            tracing::debug!(session_id = %session.id, "Restored session into local registry");
-            session
-        };
-
-        // Seed the event counter past the highest buffered event ID so new
-        // SSE events don't collide with ones the client may still replay.
-        if let Ok(events) = self.events.replay_after(&record.id, 0).await
-            && let Some(max_id) = events.iter().map(|e| e.id).max()
-        {
-            session
-                .event_counter
-                .store(max_id + 1, std::sync::atomic::Ordering::SeqCst);
-        }
-
-        // Refresh last_accessed in the store so the record doesn't expire
-        // immediately after restore.
-        let mut refreshed = record;
-        refreshed.touch(self.config.ttl);
-        if let Err(e) = self.persistent.save(&refreshed).await {
-            tracing::warn!(session_id = %refreshed.id, error = %e, "Failed to refresh restored session record");
-        }
-
-        Some(session)
-    }
-
-    /// Create a new session with the requested ID and synthetic client
-    /// info, skipping the initialize handshake. Used when `auto_reinit`
-    /// is enabled and no stored record exists.
-    ///
-    /// Loses the original client's identity and capabilities — the server
-    /// sees a session from client `"auto-recovered"`.
-    async fn auto_reinitialize(&self, id: &str) -> Option<Arc<Session>> {
-        let mut record = crate::session_store::SessionRecord::new(
-            id.to_string(),
-            LATEST_PROTOCOL_VERSION.to_string(),
-            self.config.ttl,
-        );
-        record.client_info = Some(crate::protocol::Implementation {
-            name: "auto-recovered".into(),
-            version: "unknown".into(),
-            title: None,
-            description: None,
-            icons: None,
-            website_url: None,
-            meta: None,
-        });
-        record.client_capabilities = Some(crate::protocol::ClientCapabilities::default());
-
-        // Persist first so a concurrent request sees the record. Ignore
-        // persistence errors; the in-memory session will still work.
-        if let Err(e) = self.persistent.create(&mut record).await {
-            tracing::warn!(session_id = %id, error = %e, "Failed to persist auto-reinitialized session");
-        }
-
-        self.restore_from_record(record).await
-    }
-
-    async fn remove(&self, id: &str) -> bool {
-        let removed = {
-            let mut sessions = self.sessions.write().await;
-            sessions.remove(id).is_some()
-        };
-        if removed {
-            tracing::debug!(session_id = %id, "Removed session");
-            if let Err(e) = self.persistent.delete(id).await {
-                tracing::warn!(session_id = %id, error = %e, "Failed to delete session record");
-            }
-            if let Err(e) = self.events.purge_session(id).await {
-                tracing::warn!(session_id = %id, error = %e, "Failed to purge session events");
-            }
-        }
-        removed
-    }
-
-    /// Send a pre-serialized JSON notification to every live session's SSE
-    /// broadcast channel.
-    ///
-    /// Used by the external-notification fan-out task. Failures to send
-    /// (no SSE subscribers attached to a session yet) are silent — the
-    /// broadcast channel drops the message naturally.
-    async fn broadcast_to_all(&self, json: &str) {
-        let sessions = self.sessions.read().await;
-        for session in sessions.values() {
-            let _ = session.notifications_tx.send(json.to_string());
-        }
-    }
-
-    /// Remove expired sessions, returns count of removed sessions
-    async fn cleanup_expired(&self) -> usize {
-        let expired = {
-            let mut sessions = self.sessions.write().await;
-            let ttl = self.config.ttl;
-
-            let mut expired = Vec::new();
-            for (id, session) in sessions.iter() {
-                if session.is_expired(ttl).await {
-                    expired.push(id.clone());
-                }
-            }
-
-            for id in &expired {
-                sessions.remove(id);
-                tracing::debug!(session_id = %id, "Expired session removed");
-            }
-
-            if !expired.is_empty() {
-                tracing::info!(
-                    expired_count = expired.len(),
-                    remaining = sessions.len(),
-                    "Session cleanup completed"
-                );
-            }
-            expired
-        };
-
-        for id in &expired {
-            if let Err(e) = self.persistent.delete(id).await {
-                tracing::warn!(session_id = %id, error = %e, "Failed to delete expired session record");
-            }
-            if let Err(e) = self.events.purge_session(id).await {
-                tracing::warn!(session_id = %id, error = %e, "Failed to purge expired session events");
-            }
-        }
-
-        expired.len()
-    }
-}
-
-/// Metadata about an active session.
-///
-/// Returned by [`SessionHandle::list_sessions()`].
-#[derive(Debug, Clone)]
-pub struct SessionInfo {
-    /// The session ID.
-    pub id: String,
-    /// How long ago this session was created.
-    pub created_at: Duration,
-    /// How long ago this session was last accessed.
-    pub last_activity: Duration,
-}
-
-/// A handle for managing HTTP transport sessions and final subscription streams.
-///
-/// Obtained from [`HttpTransport::into_router_with_handle()`] or
-/// [`HttpTransport::into_router_at_with_handle()`]. The handle is cheap to
-/// clone and can be shared across threads.
-///
-/// # Example
-///
-/// ```rust,ignore
-/// use tower_mcp::transport::http::HttpTransport;
-///
-/// let transport = HttpTransport::new(router);
-/// let (router, handle) = transport.into_router_with_handle();
-///
-/// // Later, in an admin endpoint:
-/// let count = handle.session_count().await;
-/// for info in handle.list_sessions().await {
-///     println!("{}: created {:?} ago", info.id, info.created_at);
-/// }
-/// handle.terminate_session("session-id").await;
-///
-/// // During graceful server shutdown (with the `stateless` feature):
-/// handle.close_subscriptions();
-/// ```
-#[derive(Clone)]
-pub struct SessionHandle {
-    store: Arc<SessionRegistry>,
-    #[cfg(feature = "stateless")]
-    modern_subscriptions: Arc<ModernSubscriptionRegistry>,
-}
-
-impl SessionHandle {
-    /// Returns the number of currently active sessions.
-    pub async fn session_count(&self) -> usize {
-        self.store.sessions.read().await.len()
-    }
-
-    /// Returns metadata for all active sessions.
-    pub async fn list_sessions(&self) -> Vec<SessionInfo> {
-        let sessions = self.store.sessions.read().await;
-        let mut infos = Vec::with_capacity(sessions.len());
-        for session in sessions.values() {
-            let last_accessed = session.last_accessed.read().await;
-            infos.push(SessionInfo {
-                id: session.id.clone(),
-                created_at: session.created_at.elapsed(),
-                last_activity: last_accessed.elapsed(),
-            });
-        }
-        infos
-    }
-
-    /// Terminates a session by ID, returning `true` if the session existed.
-    pub async fn terminate_session(&self, id: &str) -> bool {
-        self.store.remove(id).await
-    }
-
-    /// Returns the number of active final-protocol subscription streams.
-    #[cfg(feature = "stateless")]
-    pub fn subscription_count(&self) -> usize {
-        self.modern_subscriptions.len()
-    }
-
-    /// Gracefully finish every active final-protocol subscription stream.
-    ///
-    /// Each stream receives its terminal `SubscriptionsListenResult` before
-    /// closing. Returns the number of streams that were drained.
-    #[cfg(feature = "stateless")]
-    pub fn close_subscriptions(&self) -> usize {
-        self.modern_subscriptions.close_all()
-    }
-}
 
 /// The source of the MCP service for session creation.
 #[derive(Clone)]
@@ -1360,184 +377,6 @@ struct AppState {
     sse_responses: bool,
     /// Maximum accepted POST body size in bytes
     max_body_size: usize,
-}
-
-#[cfg(feature = "stateless")]
-struct ModernSubscription {
-    subscription_id: RequestId,
-    filter: SubscriptionFilter,
-    tx: mpsc::UnboundedSender<String>,
-    started: std::time::Instant,
-}
-
-/// Process-local registry for sessionless final-protocol subscriptions.
-///
-/// The 2026-07-28 transport deliberately has no session or replay state.
-/// Each listen POST owns one sender, removed when its response stream drops.
-#[cfg(feature = "stateless")]
-struct ModernSubscriptionRegistry {
-    next_key: AtomicU64,
-    subscriptions: std::sync::Mutex<HashMap<u64, ModernSubscription>>,
-    server_info: Option<Implementation>,
-    observer: Option<Arc<dyn crate::transport::subscriptions::SubscriptionObserver>>,
-}
-
-#[cfg(feature = "stateless")]
-impl ModernSubscriptionRegistry {
-    fn new(
-        server_info: Option<Implementation>,
-        observer: Option<Arc<dyn crate::transport::subscriptions::SubscriptionObserver>>,
-    ) -> Self {
-        Self {
-            next_key: AtomicU64::new(0),
-            subscriptions: std::sync::Mutex::new(HashMap::new()),
-            server_info,
-            observer,
-        }
-    }
-
-    fn observe_close(
-        &self,
-        subscription: &ModernSubscription,
-        reason: crate::transport::subscriptions::SubscriptionCloseReason,
-    ) {
-        if let Some(observer) = &self.observer {
-            observer.on_close(crate::transport::subscriptions::SubscriptionClose {
-                subscription_id: subscription.subscription_id.clone(),
-                reason,
-                duration: subscription.started.elapsed(),
-            });
-        }
-    }
-
-    fn register(
-        self: &Arc<Self>,
-        subscription_id: RequestId,
-        filter: SubscriptionFilter,
-    ) -> (mpsc::UnboundedReceiver<String>, ModernSubscriptionGuard) {
-        let key = self.next_key.fetch_add(1, Ordering::Relaxed);
-        let (tx, rx) = mpsc::unbounded_channel();
-        self.subscriptions.lock().unwrap().insert(
-            key,
-            ModernSubscription {
-                subscription_id,
-                filter,
-                tx,
-                started: std::time::Instant::now(),
-            },
-        );
-        (
-            rx,
-            ModernSubscriptionGuard {
-                key,
-                registry: self.clone(),
-            },
-        )
-    }
-
-    /// Route subscription-scoped notifications and return whether the
-    /// notification belongs exclusively on listen streams.
-    fn publish(&self, notification: &ServerNotification) -> bool {
-        let subscription_scoped = matches!(
-            notification,
-            ServerNotification::ResourceUpdated { .. }
-                | ServerNotification::ResourcesListChanged
-                | ServerNotification::ToolsListChanged
-                | ServerNotification::PromptsListChanged
-                | ServerNotification::FinalTaskStatusChanged(_)
-        );
-        if !subscription_scoped {
-            return false;
-        }
-
-        let mut subscriptions = self.subscriptions.lock().unwrap();
-        tracing::trace!(
-            active_subscriptions = subscriptions.len(),
-            notification = ?notification,
-            "Routing final-protocol subscription notification"
-        );
-        subscriptions.retain(|_, subscription| {
-            let keep = if subscription_matches(notification, &subscription.filter)
-                && let Some(json) =
-                    tagged_subscription_notification(notification, &subscription.subscription_id)
-            {
-                subscription.tx.send(json).is_ok()
-            } else {
-                !subscription.tx.is_closed()
-            };
-            if !keep {
-                // Detected here rather than at guard drop: the entry is
-                // removed exactly once, so the close is reported exactly once.
-                self.observe_close(
-                    subscription,
-                    crate::transport::subscriptions::SubscriptionCloseReason::Disconnected,
-                );
-            }
-            keep
-        });
-        true
-    }
-
-    fn len(&self) -> usize {
-        self.subscriptions.lock().unwrap().len()
-    }
-
-    /// Gracefully finish every active HTTP listen stream.
-    fn close_all(&self) -> usize {
-        let subscriptions = {
-            let mut active = self.subscriptions.lock().unwrap();
-            active
-                .drain()
-                .map(|(_, subscription)| subscription)
-                .collect::<Vec<_>>()
-        };
-        let count = subscriptions.len();
-        for subscription in subscriptions {
-            self.observe_close(
-                &subscription,
-                crate::transport::subscriptions::SubscriptionCloseReason::Drained,
-            );
-            let response = subscription_complete_response(
-                subscription.subscription_id,
-                self.server_info.clone(),
-            );
-            if let Ok(json) = serde_json::to_string(&response) {
-                let _ = subscription.tx.send(json);
-            }
-        }
-        count
-    }
-}
-
-#[cfg(feature = "stateless")]
-impl Default for ModernSubscriptionRegistry {
-    fn default() -> Self {
-        Self::new(None, None)
-    }
-}
-
-#[cfg(feature = "stateless")]
-struct ModernSubscriptionGuard {
-    key: u64,
-    registry: Arc<ModernSubscriptionRegistry>,
-}
-
-#[cfg(feature = "stateless")]
-impl Drop for ModernSubscriptionGuard {
-    fn drop(&mut self) {
-        let removed = self
-            .registry
-            .subscriptions
-            .lock()
-            .unwrap()
-            .remove(&self.key);
-        if let Some(subscription) = removed {
-            self.registry.observe_close(
-                &subscription,
-                crate::transport::subscriptions::SubscriptionCloseReason::Disconnected,
-            );
-        }
-    }
 }
 
 /// Configuration for OAuth 2.1 Protected Resource Metadata.
@@ -2652,2218 +1491,27 @@ fn spawn_external_notification_fanout(
     });
 }
 
-fn is_localhost_origin(origin: &str) -> bool {
-    // Parse the origin to extract the host
-    if let Some(rest) = origin
-        .strip_prefix("http://")
-        .or_else(|| origin.strip_prefix("https://"))
-    {
-        is_localhost_host(rest)
-    } else {
-        false
-    }
-}
-
-/// Check if a `host:port` (or `[ipv6]:port`) value refers to localhost.
-///
-/// Used by both Origin validation (after stripping the `http(s)://` scheme)
-/// and Host validation (where there's no scheme to begin with).
-fn is_localhost_host(host: &str) -> bool {
-    let host_only = if host.starts_with('[') {
-        // Bracketed IPv6: [::1]:3000 -> ::1
-        host.split(']')
-            .next()
-            .unwrap_or(host)
-            .trim_start_matches('[')
-    } else {
-        // Strip port if present
-        host.split(':').next().unwrap_or(host)
-    };
-    matches!(host_only, "localhost" | "127.0.0.1" | "::1")
-}
-
-/// Resolve the effective host for validation.
-///
-/// Prefers the `Host` header, falling back to the HTTP/2 `:authority`
-/// pseudo-header (`request.uri().authority()`) when the header is missing.
-/// This matters behind middleware like `axum::Router::nest`, which can
-/// strip Hyper's synthesized `Host` before our handler sees it.
-fn effective_host<'a>(headers: &'a HeaderMap, uri: &'a axum::http::Uri) -> Option<&'a str> {
-    if let Some(value) = headers.get(header::HOST)
-        && let Ok(s) = value.to_str()
-    {
-        return Some(s);
-    }
-    uri.authority().map(|a| a.as_str())
-}
-
-/// Validate the `Host` header (defense-in-depth alongside Origin).
-///
-/// Returns Some(Response) if validation fails, None if it passes.
-fn validate_host(headers: &HeaderMap, uri: &axum::http::Uri, state: &AppState) -> Option<Response> {
-    if !state.validate_host {
-        return None;
-    }
-
-    let Some(host) = effective_host(headers, uri) else {
-        if state.allowed_hosts.is_empty() {
-            // No Host header and no allowlist: fall back to permissive
-            // behavior matching pre-validation defaults so we don't break
-            // existing deployments. (Origin already protects browsers.)
-            return None;
-        }
-        tracing::warn!("Rejecting request: missing Host header and no :authority fallback");
-        return Some((StatusCode::BAD_REQUEST, "Missing Host header").into_response());
-    };
-
-    if is_localhost_host(host) {
-        return None;
-    }
-
-    if state.allowed_hosts.is_empty() {
-        // Non-localhost host with no explicit allowlist: keep accepting it.
-        // Operators who want strict Host validation must opt in via
-        // `.allowed_hosts(...)`. This preserves the historical behavior of
-        // not enforcing Host on non-loopback deployments by default.
-        return None;
-    }
-
-    if state.allowed_hosts.iter().any(|h| h == host) {
-        return None;
-    }
-
-    tracing::warn!(host = %host, "Rejecting request: Host not in allowlist");
-    Some((StatusCode::BAD_REQUEST, "Host not allowed").into_response())
-}
-
-/// Validate Origin header for security.
-///
-/// When origin validation is enabled:
-/// - Requests without an Origin header are allowed (same-origin)
-/// - Localhost origins are always allowed (DNS rebinding protection)
-/// - If `allowed_origins` is non-empty, non-localhost origins must match
-/// - If `allowed_origins` is empty, non-localhost origins are rejected
-///
-/// Returns Some(Response) if validation fails, None if it passes.
-fn validate_origin(headers: &HeaderMap, state: &AppState) -> Option<Response> {
-    if !state.validate_origin {
-        return None;
-    }
-
-    if let Some(origin) = headers.get(header::ORIGIN) {
-        let origin_str = origin.to_str().unwrap_or("");
-
-        // Always allow localhost origins (DNS rebinding protection allows these)
-        if is_localhost_origin(origin_str) {
-            return None;
-        }
-
-        // Non-localhost origin: check against allowed list
-        if state.allowed_origins.is_empty() {
-            tracing::warn!(
-                origin = %origin_str,
-                "Rejecting request: cross-origin not allowed (no allowlist configured)"
-            );
-            return Some(
-                (StatusCode::FORBIDDEN, "Cross-origin requests not allowed").into_response(),
-            );
-        }
-
-        if !state
-            .allowed_origins
-            .iter()
-            .any(|o| o == origin_str || o == "*")
-        {
-            tracing::warn!(origin = %origin_str, "Rejecting request: Origin not in allowlist");
-            return Some((StatusCode::FORBIDDEN, "Origin not allowed").into_response());
-        }
-    }
-
-    None
-}
-
-/// Extract and validate session ID from headers
-fn get_session_id(headers: &HeaderMap) -> Option<String> {
-    headers
-        .get(MCP_SESSION_ID_HEADER)
-        .and_then(|v| v.to_str().ok())
-        .map(|s| s.to_string())
-}
-
-/// Extract protocol version from headers
-fn get_protocol_version(headers: &HeaderMap) -> Option<String> {
-    headers
-        .get(MCP_PROTOCOL_VERSION_HEADER)
-        .and_then(|v| v.to_str().ok())
-        .map(|s| s.to_string())
-}
-
-/// Extract Last-Event-ID from headers for SSE stream resumption (SEP-1699)
-fn get_last_event_id(headers: &HeaderMap) -> Option<u64> {
-    headers
-        .get(LAST_EVENT_ID_HEADER)
-        .and_then(|v| v.to_str().ok())
-        .and_then(|s| s.parse::<u64>().ok())
-}
-
-/// Check if the request is an initialize request
-fn is_initialize_request(body: &serde_json::Value) -> bool {
-    body.get("method")
-        .and_then(|m| m.as_str())
-        .map(|m| m == "initialize")
-        .unwrap_or(false)
-}
-
-/// Check if this is a response to one of our outgoing requests
-fn is_response(parsed: &serde_json::Value) -> bool {
-    parsed.get("method").is_none()
-        && (parsed.get("result").is_some() || parsed.get("error").is_some())
-}
-
-/// Resolve the selected tool's input schema when the transport owns an
-/// [`McpRouter`]. Pre-built services do not expose their tool registry, so
-/// supplied custom headers can still be checked there but missing headers
-/// cannot be inferred before dispatch.
-fn request_tool_input_schema(
-    service_source: &ServiceSource,
-    parsed: &serde_json::Value,
-) -> Option<serde_json::Value> {
-    if parsed.get("method").and_then(serde_json::Value::as_str) != Some("tools/call") {
-        return None;
-    }
-    let name = parsed
-        .get("params")
-        .and_then(serde_json::Value::as_object)
-        .and_then(|params| params.get("name"))
-        .and_then(serde_json::Value::as_str)?;
-    match service_source {
-        ServiceSource::Router { router, .. } => router.tool_input_schema(name),
-        ServiceSource::Service(_) => None,
-    }
-}
-
-/// Return whether an HTTP request claims the modern, per-request-metadata
-/// protocol era.
-///
-/// The body envelope is authoritative for era detection. The final-version
-/// header is also treated as a modern claim so a missing or malformed
-/// envelope receives the specified modern error instead of drifting into the
-/// legacy session path.
-fn claims_modern_protocol(headers: &HeaderMap, parsed: &serde_json::Value) -> bool {
-    get_protocol_version(headers).as_deref() == Some(PROTOCOL_VERSION_2026_07_28)
-        || parsed
-            .get("params")
-            .and_then(serde_json::Value::as_object)
-            .and_then(|params| params.get("_meta"))
-            .and_then(serde_json::Value::as_object)
-            .is_some_and(|meta| meta.contains_key("io.modelcontextprotocol/protocolVersion"))
-}
-
-/// Validate the required modern per-request metadata and return its declared
-/// protocol version.
-///
-/// `clientInfo` is deliberately optional in the final specification.
-fn validate_modern_request_meta(
-    parsed: &serde_json::Value,
-) -> std::result::Result<String, JsonRpcError> {
-    let params = parsed
-        .get("params")
-        .and_then(serde_json::Value::as_object)
-        .ok_or_else(|| {
-            JsonRpcError::invalid_params("Modern requests require a params object containing _meta")
-        })?;
-    let meta_value = params
-        .get("_meta")
-        .ok_or_else(|| JsonRpcError::invalid_params("Modern requests require a _meta object"))?;
-    crate::protocol::validate_meta_object(meta_value)
-        .map_err(|error| JsonRpcError::invalid_params(error.to_string()))?;
-    let meta = meta_value
-        .as_object()
-        .expect("validate_meta_object accepted a JSON object");
-    let protocol_version = meta
-        .get("io.modelcontextprotocol/protocolVersion")
-        .and_then(serde_json::Value::as_str)
-        .ok_or_else(|| {
-            JsonRpcError::invalid_params(
-                "Missing or invalid _meta.io.modelcontextprotocol/protocolVersion",
-            )
-        })?;
-    let client_capabilities = meta
-        .get("io.modelcontextprotocol/clientCapabilities")
-        .ok_or_else(|| {
-            JsonRpcError::invalid_params("Missing _meta.io.modelcontextprotocol/clientCapabilities")
-        })?;
-    if !client_capabilities.is_object()
-        || serde_json::from_value::<ClientCapabilities>(client_capabilities.clone()).is_err()
-    {
-        return Err(JsonRpcError::invalid_params(
-            "Invalid _meta.io.modelcontextprotocol/clientCapabilities",
-        ));
-    }
-
-    Ok(protocol_version.to_string())
-}
-
-/// Methods present in legacy protocol unions but removed from the modern core.
-fn is_removed_modern_method(method: &str) -> bool {
-    matches!(
-        method,
-        "initialize"
-            | "notifications/initialized"
-            | "ping"
-            | "logging/setLevel"
-            | "resources/subscribe"
-            | "resources/unsubscribe"
-            | "notifications/roots/list_changed"
-    )
-}
-
-/// Map protocol errors whose final Streamable HTTP binding assigns a
-/// non-success status. Errors emitted after an SSE stream has opened remain
-/// in-band because the HTTP status is already committed.
+mod handlers;
+mod session;
 #[cfg(feature = "stateless")]
-fn modern_response_status(response: &JsonRpcResponse) -> StatusCode {
-    let JsonRpcResponse::Error(error) = response else {
-        return StatusCode::OK;
-    };
-    if error.error.code == ErrorCode::MethodNotFound as i32 {
-        StatusCode::NOT_FOUND
-    } else if error.error.code == McpErrorCode::MissingRequiredClientCapability.code() {
-        StatusCode::BAD_REQUEST
-    } else {
-        StatusCode::OK
-    }
-}
-
-/// Extract request ID from a JSON value
-fn extract_request_id(parsed: &serde_json::Value) -> Option<RequestId> {
-    parsed.get("id").and_then(|id| {
-        if let Some(n) = id.as_i64() {
-            Some(RequestId::Number(n))
-        } else {
-            id.as_str().map(|s| RequestId::String(s.to_string()))
-        }
-    })
-}
-
-/// Handle POST requests (JSON-RPC messages from client)
-async fn handle_post(
-    State(state): State<Arc<AppState>>,
-    request: axum::extract::Request,
-) -> Response {
-    let (parts, body_bytes) = request.into_parts();
-    let headers = parts.headers;
-    let uri = parts.uri.clone();
-
-    // Validate Host (DNS rebinding defense, complement to Origin)
-    if let Some(resp) = validate_host(&headers, &uri, &state) {
-        return resp;
-    }
-
-    // Validate Origin
-    if let Some(resp) = validate_origin(&headers, &state) {
-        return resp;
-    }
-
-    // Bound the body size (rmcp #970 analog). axum's `DefaultBodyLimit`
-    // doesn't apply here because this handler consumes the raw `Request`
-    // instead of a body-consuming extractor, so this is the only limit on
-    // the MCP POST body. A declared Content-Length above the limit is
-    // rejected without reading; chunked bodies are capped while streaming.
-    if let Some(declared) = headers
-        .get(header::CONTENT_LENGTH)
-        .and_then(|v| v.to_str().ok())
-        .and_then(|v| v.parse::<usize>().ok())
-        && declared > state.max_body_size
-    {
-        return body_too_large_response(state.max_body_size);
-    }
-
-    let body = match axum::body::to_bytes(body_bytes, state.max_body_size).await {
-        Ok(bytes) => match String::from_utf8(bytes.to_vec()) {
-            Ok(s) => s,
-            Err(e) => {
-                return json_rpc_error_response(
-                    None,
-                    JsonRpcError::parse_error(format!("Invalid UTF-8: {}", e)),
-                );
-            }
-        },
-        Err(e) if is_length_limit_error(&e) => {
-            return body_too_large_response(state.max_body_size);
-        }
-        Err(e) => {
-            return json_rpc_error_response(
-                None,
-                JsonRpcError::parse_error(format!("Failed to read body: {}", e)),
-            );
-        }
-    };
-
-    // Per-request data bridged from HTTP into MCP extensions: OAuth claims
-    // when that feature is compiled in, plus whatever types the server
-    // registered with `bridge_extension`, which is independent of OAuth
-    // (#1242). Always bound, since the bridges run in every build.
-    let http_extensions = parts.extensions;
-
-    // Parse the request body
-    let parsed: serde_json::Value = match serde_json::from_str(&body) {
-        Ok(v) => v,
-        Err(e) => {
-            return json_rpc_error_response(
-                None,
-                JsonRpcError::parse_error(format!("Invalid JSON: {}", e)),
-            );
-        }
-    };
-
-    // A version header supplies enough exact context to reject a batch before
-    // any object-only HTTP classification runs. Legacy batches without a
-    // header are validated against their session revision after lookup below.
-    if parsed.is_array()
-        && let Some(version) = get_protocol_version(&headers)
-    {
-        let revision = match version.parse::<McpProtocolRevision>() {
-            Ok(revision) => revision,
-            Err(_) => {
-                return json_rpc_error_response(
-                    None,
-                    JsonRpcError::unsupported_protocol_version(
-                        version,
-                        state.protocol_support.versions().iter().map(String::as_str),
-                    ),
-                );
-            }
-        };
-        if let Err(error) = inspect_runtime_value(
-            &parsed,
-            revision,
-            &state.protocol_support,
-            McpDirection::ClientToServer,
-        ) {
-            let status = if revision == McpProtocolRevision::V2026_07_28 {
-                StatusCode::BAD_REQUEST
-            } else {
-                StatusCode::OK
-            };
-            return json_rpc_error_response_with_status(None, error, status);
-        }
-    }
-
-    // Check if this is an initialize request (creates new session)
-    let is_init = is_initialize_request(&parsed);
-    let request_method = parsed
-        .get("method")
-        .and_then(|method| method.as_str())
-        .unwrap_or_default()
-        .to_string();
-    let tool_input_schema = request_tool_input_schema(&state.service_source, &parsed);
-    let modern_request = claims_modern_protocol(&headers, &parsed);
-
-    // The modern protocol is selected by its per-request `_meta` envelope,
-    // with the final-version HTTP header also acting as a signal for malformed
-    // requests whose envelope is missing. Resolve that era before consulting
-    // any legacy session state so modern traffic cannot accidentally fall
-    // through to the initialize/session lifecycle.
-    if modern_request {
-        let id = extract_request_id(&parsed);
-        let body_version = match validate_modern_request_meta(&parsed) {
-            Ok(version) => version,
-            Err(error) => {
-                return json_rpc_error_response_with_status(id, error, StatusCode::BAD_REQUEST);
-            }
-        };
-
-        let Some(header_version) = get_protocol_version(&headers) else {
-            return json_rpc_error_response_with_status(
-                id,
-                JsonRpcError::header_mismatch("MCP-Protocol-Version header is required"),
-                StatusCode::BAD_REQUEST,
-            );
-        };
-        if header_version != body_version {
-            return json_rpc_error_response_with_status(
-                id,
-                JsonRpcError::header_mismatch(format!(
-                    "MCP-Protocol-Version header value {header_version:?} does not match \
-                     request _meta protocol version {body_version:?}"
-                )),
-                StatusCode::BAD_REQUEST,
-            );
-        }
-
-        if !state.protocol_support.contains(&body_version) {
-            return json_rpc_error_response_with_status(
-                id,
-                JsonRpcError::unsupported_protocol_version(
-                    body_version,
-                    state.protocol_support.versions().iter().map(String::as_str),
-                ),
-                StatusCode::BAD_REQUEST,
-            );
-        }
-
-        let revision = match body_version.parse::<McpProtocolRevision>() {
-            Ok(revision) => revision,
-            Err(_) => {
-                return json_rpc_error_response_with_status(
-                    id,
-                    JsonRpcError::unsupported_protocol_version(
-                        body_version,
-                        state.protocol_support.versions().iter().map(String::as_str),
-                    ),
-                    StatusCode::BAD_REQUEST,
-                );
-            }
-        };
-        if let Err(error) = inspect_runtime_value(
-            &parsed,
-            revision,
-            &state.protocol_support,
-            McpDirection::ClientToServer,
-        ) {
-            return json_rpc_error_response_with_status(id, error, StatusCode::BAD_REQUEST);
-        }
-
-        let sep_2243_mode = super::http_headers::mode_for_version(&body_version);
-        if let Err(error) = super::http_headers::validate_with_tool_schema(
-            &headers,
-            &parsed,
-            sep_2243_mode,
-            tool_input_schema.as_ref(),
-        ) {
-            tracing::warn!(
-                mode = ?sep_2243_mode,
-                version = %body_version,
-                error = %error.message,
-                "Rejecting modern request: HTTP header validation failed",
-            );
-            return json_rpc_error_response_with_status(id, error, StatusCode::BAD_REQUEST);
-        }
-
-        if is_removed_modern_method(&request_method) {
-            return json_rpc_error_response_with_status(
-                id,
-                JsonRpcError::method_not_found(&request_method),
-                StatusCode::NOT_FOUND,
-            );
-        }
-    }
-
-    // SEP-2575 / SEP-2567: version-gated stateless mode for 2026-07-28+ clients.
-    //
-    // When the requested (or carried) protocol version is >= 2026-07-28 and the
-    // request has no mcp-session-id, every request -- including `initialize` --
-    // is served without creating or looking up a session. Each request is fully
-    // self-contained; client identity and capabilities flow through per-request
-    // `_meta` rather than a session handshake.
-    //
-    // This block runs before the legacy SEP-1442 stateless path so that
-    // 2026-07-28 requests are handled here regardless of whether
-    // `stateless_config` is set on the transport.
-    #[cfg(feature = "stateless")]
-    {
-        let version_in_play: Option<String> = if is_init && !modern_request {
-            // For `initialize`, read the version the client is requesting from
-            // the params object.
-            parsed
-                .get("params")
-                .and_then(|p| p.get("protocolVersion"))
-                .and_then(|v| v.as_str())
-                .map(|s| s.to_string())
-        } else {
-            // For non-init requests, only the HTTP-level `MCP-Protocol-Version`
-            // header gates stateless mode. Body-level `_meta.protocolVersion` is
-            // plumbed to handlers via `stash_per_request_meta` in both paths.
-            get_protocol_version(&headers)
-        };
-
-        if let Some(ref version) = version_in_play
-            && is_stateless_protocol_version(version)
-            && state.protocol_support.contains(version)
-            // `subscriptions/listen` opens an SSE stream; let it fall through to the
-            // dedicated intercept below rather than handling it as a plain RPC call.
-            && parsed.get("method").and_then(|m| m.as_str()) != Some("subscriptions/listen")
-        {
-            // Notifications and responses are fire-and-forget; no dispatch needed.
-            if !is_init && (parsed.get("id").is_none() || is_response(&parsed)) {
-                return StatusCode::ACCEPTED.into_response();
-            }
-
-            // SEP-2243 validation before `parsed` is consumed by deserialization.
-            // 2026-07-28 falls into strict mode, so missing Mcp-Method is an error.
-            let sep_2243_mode = super::http_headers::mode_for_version(version);
-            if let Err(err) = super::http_headers::validate_with_tool_schema(
-                &headers,
-                &parsed,
-                sep_2243_mode,
-                tool_input_schema.as_ref(),
-            ) {
-                tracing::warn!(
-                    mode = ?sep_2243_mode,
-                    version = %version,
-                    error = %err.message,
-                    "Rejecting stateless request: SEP-2243 header validation failed",
-                );
-                let id = extract_request_id(&parsed);
-                let mut resp = json_rpc_error_response(id, err);
-                *resp.status_mut() = StatusCode::BAD_REQUEST;
-                return resp;
-            }
-
-            let request: JsonRpcRequest = match serde_json::from_value(parsed) {
-                Ok(r) => r,
-                Err(e) => {
-                    return json_rpc_error_response(
-                        None,
-                        JsonRpcError::parse_error(format!("Invalid request: {}", e)),
-                    );
-                }
-            };
-
-            // Ephemeral pre-initialized service -- no session is stored or created.
-            //
-            // A per-request notification channel captures anything the handler
-            // emits during the call (progress, logging). With no session and no
-            // GET stream on this path, those messages can only reach the client
-            // on the POST response itself: per the draft Streamable HTTP rules,
-            // a plain JSON body is only correct when the first outbound message
-            // is the terminal response; otherwise the response falls back to
-            // SSE with the notifications delivered ahead of the terminal
-            // response.
-            // Captured before the match below borrows `router` into the
-            // ephemeral session; used to stamp `_meta.serverInfo` on the
-            // outgoing response (SEP-2575). `None` for a transport built
-            // from a pre-built service (no router to read identity from).
-            let server_identity = match &state.service_source {
-                ServiceSource::Router { router, .. } if state.stamp_server_info => {
-                    Some(router.implementation())
-                }
-                _ => None,
-            };
-
-            let (notif_tx, mut notif_rx) = crate::context::notification_channel(64);
-            let mut service = match &state.service_source {
-                ServiceSource::Router { router, factory } => {
-                    let ephemeral = router
-                        .with_fresh_session()
-                        .with_request_notification_sender(notif_tx);
-                    ephemeral.session().mark_preinitialized();
-                    JsonRpcService::new(factory(ephemeral))
-                }
-                ServiceSource::Service(mutex) => JsonRpcService::new(mutex.lock().unwrap().clone()),
-            };
-
-            let mut ext = crate::router::Extensions::new();
-            ext.insert(state.protocol_support.clone());
-            #[cfg(feature = "oauth")]
-            if let Some(claims) = http_extensions.get::<crate::oauth::token::TokenClaims>() {
-                ext.insert(claims.clone());
-            }
-            stash_per_request_meta(&request, &mut ext);
-            crate::transport::extension_bridge::apply_extension_bridges(
-                &state.extension_bridges,
-                &http_extensions,
-                &mut ext,
-            );
-
-            // rmcp #967 analog: give the request a cancellation token that
-            // fires if the client disconnects before the response is
-            // delivered. The router adopts the token as the
-            // `RequestContext`'s cancellation source, so handlers observe
-            // the disconnect via `ctx.is_cancelled()` / `ctx.cancelled()`,
-            // and spawned work holding a token clone is signalled even
-            // after the request future itself is dropped. Session-based
-            // requests are exempt: with stream resumption, a disconnect is
-            // not a cancellation.
-            let cancel_token = crate::context::CancellationToken::new();
-            let mut cancel_guard = CancelOnDisconnect::arm(cancel_token.clone());
-            ext.insert(cancel_token);
-
-            service = service.with_extensions(ext);
-
-            let mut call: std::pin::Pin<
-                Box<dyn std::future::Future<Output = crate::error::Result<JsonRpcResponse>> + Send>,
-            > = Box::pin(async move {
-                let mut service = service;
-                service.call_single(request).await
-            });
-
-            enum FirstOutbound {
-                Response(crate::error::Result<JsonRpcResponse>),
-                Notification(crate::context::ServerNotification),
-            }
-
-            // Race the handler against its first notification. A closed
-            // channel (no sender attached, or all senders dropped) simply
-            // awaits the handler.
-            let first = loop {
-                let outbound = tokio::select! {
-                    // A handler may enqueue a notification and complete in
-                    // the same poll. Observe the queued notification first so
-                    // it is neither dropped nor raced behind the response.
-                    biased;
-                    maybe = notif_rx.recv() => match maybe {
-                        Some(n) => FirstOutbound::Notification(n),
-                        None => FirstOutbound::Response((&mut call).await),
-                    },
-                    result = &mut call => FirstOutbound::Response(result),
-                };
-                match outbound {
-                    FirstOutbound::Notification(notification)
-                        if state.modern_subscriptions.publish(&notification) =>
-                    {
-                        continue;
-                    }
-                    outbound => break outbound,
-                }
-            };
-
-            match first {
-                FirstOutbound::Response(result) => {
-                    // `select!` may observe a handler's ready response in the
-                    // same poll that the handler enqueued notifications.
-                    // Drain that queue before committing a JSON response.
-                    while let Ok(notification) = notif_rx.try_recv() {
-                        if state.modern_subscriptions.publish(&notification) {
-                            continue;
-                        }
-                        let ready_call: std::pin::Pin<
-                            Box<
-                                dyn std::future::Future<
-                                        Output = crate::error::Result<JsonRpcResponse>,
-                                    > + Send,
-                            >,
-                        > = Box::pin(async move { result });
-                        let mut resp = stateless_sse_with_notifications(
-                            notification,
-                            ready_call,
-                            notif_rx,
-                            StatelessSseContext {
-                                version: version.clone(),
-                                method: request_method.clone(),
-                                cancel_guard,
-                                server_identity,
-                                subscriptions: state.modern_subscriptions.clone(),
-                            },
-                        );
-                        resp.headers_mut().insert(
-                            MCP_PROTOCOL_VERSION_HEADER,
-                            HeaderValue::from_str(version).unwrap(),
-                        );
-                        return resp;
-                    }
-
-                    // Handler finished; the response is about to be
-                    // produced, so dropping the connection from here on is
-                    // no longer a cancellation.
-                    cancel_guard.disarm();
-                    let mut response = match result {
-                        Ok(resp) => resp,
-                        Err(e) => {
-                            return json_rpc_error_response(
-                                None,
-                                JsonRpcError::internal_error(e.to_string()),
-                            );
-                        }
-                    };
-
-                    // Keep the response aligned with the version selected for
-                    // this sessionless request. The router also receives the
-                    // runtime allow-list through Extensions.
-                    if is_init
-                        && let JsonRpcResponse::Result(ref mut result) = response
-                        && let Some(pv) = result.result.get_mut("protocolVersion")
-                    {
-                        *pv = serde_json::Value::String(version.clone());
-                    }
-                    apply_protocol_result_fields(&mut response, &request_method, version);
-                    if let Some(ref identity) = server_identity {
-                        stamp_server_info(&mut response, identity);
-                    }
-
-                    let status = modern_response_status(&response);
-                    let mut resp = if state.sse_responses {
-                        sse_json_response(&response)
-                    } else {
-                        axum::Json(response).into_response()
-                    };
-                    *resp.status_mut() = status;
-                    resp.headers_mut().insert(
-                        MCP_PROTOCOL_VERSION_HEADER,
-                        HeaderValue::from_str(version).unwrap(),
-                    );
-                    // Intentionally NO `mcp-session-id` header for 2026-07-28+ clients.
-                    return resp;
-                }
-                FirstOutbound::Notification(first_notif) => {
-                    let mut resp = stateless_sse_with_notifications(
-                        first_notif,
-                        call,
-                        notif_rx,
-                        StatelessSseContext {
-                            version: version.clone(),
-                            method: request_method.clone(),
-                            cancel_guard,
-                            server_identity,
-                            subscriptions: state.modern_subscriptions.clone(),
-                        },
-                    );
-                    resp.headers_mut().insert(
-                        MCP_PROTOCOL_VERSION_HEADER,
-                        HeaderValue::from_str(version).unwrap(),
-                    );
-                    // Intentionally NO `mcp-session-id` header for 2026-07-28+ clients.
-                    return resp;
-                }
-            }
-        }
-    }
-
-    // SEP-1442: Handle stateless requests (no session needed).
-    // Stateless requests have a protocol version but no session ID and are not
-    // initialize requests. They are processed with an ephemeral service and
-    // return immediately without storing any session state.
-    #[cfg(feature = "stateless")]
-    if !is_init && state.stateless_config.is_some() && get_session_id(&headers).is_none() {
-        let version_from_header = get_protocol_version(&headers);
-        let params = parsed.get("params").unwrap_or(&parsed);
-        let version_from_meta = crate::stateless::StatelessRequestMeta::from_params(params)
-            .and_then(|m| m.protocol_version);
-
-        if let Some(version) = version_from_header.or(version_from_meta) {
-            if let Err(err) = crate::stateless::validate_protocol_version(&version) {
-                return json_rpc_error_response(None, err);
-            }
-
-            // Notifications and responses don't make sense without a session
-            if parsed.get("id").is_none() || is_response(&parsed) {
-                return StatusCode::ACCEPTED.into_response();
-            }
-
-            let request: JsonRpcRequest = match serde_json::from_value(parsed) {
-                Ok(r) => r,
-                Err(e) => {
-                    return json_rpc_error_response(
-                        None,
-                        JsonRpcError::parse_error(format!("Invalid request: {}", e)),
-                    );
-                }
-            };
-
-            // Ephemeral pre-initialized service -- no session stored
-            let mut service = match &state.service_source {
-                ServiceSource::Router { router, factory } => {
-                    let ephemeral = router.with_fresh_session();
-                    ephemeral.session().mark_preinitialized();
-                    JsonRpcService::new(factory(ephemeral))
-                }
-                ServiceSource::Service(mutex) => JsonRpcService::new(mutex.lock().unwrap().clone()),
-            };
-
-            let mut ext = crate::router::Extensions::new();
-            ext.insert(state.protocol_support.clone());
-            #[cfg(feature = "oauth")]
-            if let Some(claims) = http_extensions.get::<crate::oauth::token::TokenClaims>() {
-                ext.insert(claims.clone());
-            }
-            #[cfg(feature = "stateless")]
-            stash_per_request_meta(&request, &mut ext);
-            crate::transport::extension_bridge::apply_extension_bridges(
-                &state.extension_bridges,
-                &http_extensions,
-                &mut ext,
-            );
-            if !ext.is_empty() {
-                service = service.with_extensions(ext);
-            }
-
-            let mut response = match service.call_single(request).await {
-                Ok(resp) => resp,
-                Err(e) => {
-                    return json_rpc_error_response(
-                        None,
-                        JsonRpcError::internal_error(e.to_string()),
-                    );
-                }
-            };
-            apply_protocol_result_fields(&mut response, &request_method, &version);
-
-            let mut resp = if state.sse_responses {
-                sse_json_response(&response)
-            } else {
-                axum::Json(response).into_response()
-            };
-            resp.headers_mut().insert(
-                MCP_PROTOCOL_VERSION_HEADER,
-                HeaderValue::from_str(&version).unwrap(),
-            );
-            return resp;
-        }
-    }
-
-    // Final-protocol subscriptions are sessionless long-lived POSTs. They
-    // must be established before consulting any legacy session state.
-    #[cfg(feature = "stateless")]
-    if modern_request && request_method == "subscriptions/listen" {
-        return handle_modern_subscriptions_listen_sse(state, &parsed, &http_extensions).await;
-    }
-
-    // Runtime allowlist enforcement precedes semantic profile validation.
-    // This is especially important for optional-session traffic: an unknown
-    // header must not be interpreted under a fallback revision.
-    if !is_init
-        && let Some(version) = get_protocol_version(&headers)
-        && !state.protocol_support.contains(&version)
-    {
-        return json_rpc_error_response(
-            extract_request_id(&parsed),
-            JsonRpcError::unsupported_protocol_version(
-                version,
-                state.protocol_support.versions().iter().map(String::as_str),
-            ),
-        );
-    }
-
-    let uses_transient_session = !is_init
-        && !modern_request
-        && get_session_id(&headers).is_none()
-        && state.optional_sessions;
-
-    // Get or create session
-    let session = if is_init {
-        // Create new session for initialize
-        let create_result = match &state.service_source {
-            ServiceSource::Router { router, factory } => {
-                // Use with_fresh_session() to ensure each session has its own state
-                state
-                    .sessions
-                    .create(router.with_fresh_session(), factory.clone())
-                    .await
-            }
-            ServiceSource::Service(mutex) => {
-                let service = mutex.lock().unwrap().clone();
-                state.sessions.create_from_service(service).await
-            }
-        };
-        match create_result {
-            Some(s) => s,
-            None => {
-                return (
-                    StatusCode::SERVICE_UNAVAILABLE,
-                    "Maximum session limit reached",
-                )
-                    .into_response();
-            }
-        }
-    } else if !modern_request && let Some(session_id) = get_session_id(&headers) {
-        // Client sent a session ID -- look it up
-        match state.sessions.get(&session_id).await {
-            Some(s) => s,
-            None => {
-                // Return JSON-RPC error with session info so clients know to re-initialize
-                return json_rpc_error_response(
-                    None,
-                    JsonRpcError::session_not_found_with_id(&session_id),
-                );
-            }
-        }
-    } else if state.optional_sessions {
-        // No session ID, but sessions are optional -- create a transient,
-        // pre-initialized session so the router won't reject the request.
-        // This supports clients (Codex CLI, Cursor, etc.) that perform
-        // initialize + tools/list during setup but don't carry the session
-        // ID forward to subsequent requests.
-        let create_result = match &state.service_source {
-            ServiceSource::Router { router, factory } => {
-                state
-                    .sessions
-                    .create_initialized(router.with_fresh_session(), factory.clone())
-                    .await
-            }
-            ServiceSource::Service(mutex) => {
-                let service = mutex.lock().unwrap().clone();
-                state
-                    .sessions
-                    .create_initialized_from_service(service)
-                    .await
-            }
-        };
-        match create_result {
-            Some(s) => s,
-            None => {
-                return (
-                    StatusCode::SERVICE_UNAVAILABLE,
-                    "Maximum session limit reached",
-                )
-                    .into_response();
-            }
-        }
-    } else {
-        // No session ID and sessions are required
-        return json_rpc_error_response(None, JsonRpcError::session_required());
-    };
-
-    // Session lookup establishes the exact legacy revision. Validate the raw
-    // envelope before object-only notification/response routing, then let the
-    // existing request dispatcher consume the typed shape.
-    let session_protocol_version = if uses_transient_session {
-        let version = state
-            .protocol_support
-            .versions()
-            .iter()
-            .find(|version| {
-                crate::protocol::SUPPORTED_PROTOCOL_VERSIONS.contains(&version.as_str())
-            })
-            .map_or_else(
-                || state.protocol_support.preferred().to_string(),
-                Clone::clone,
-            );
-        *session.protocol_version.write().await = version.clone();
-        version
-    } else {
-        session.protocol_version.read().await.clone()
-    };
-    let session_revision = match session_protocol_version.parse::<McpProtocolRevision>() {
-        Ok(revision) => revision,
-        Err(_) => {
-            return json_rpc_error_response(
-                extract_request_id(&parsed),
-                JsonRpcError::unsupported_protocol_version(
-                    session_protocol_version,
-                    state.protocol_support.versions().iter().map(String::as_str),
-                ),
-            );
-        }
-    };
-    if !is_init
-        && let Err(error) = inspect_runtime_value(
-            &parsed,
-            session_revision,
-            &state.protocol_support,
-            McpDirection::ClientToServer,
-        )
-    {
-        return json_rpc_error_response(extract_request_id(&parsed), error);
-    }
-
-    if parsed.is_array() {
-        if state.strict_initialization
-            && !session
-                .initialized_notification_received
-                .load(Ordering::Acquire)
-        {
-            return json_rpc_error_response(
-                None,
-                JsonRpcError::invalid_request(
-                    "Client must send notifications/initialized before making requests",
-                ),
-            );
-        }
-
-        let message: JsonRpcMessage = match serde_json::from_value(parsed) {
-            Ok(message) => message,
-            Err(error) => {
-                return json_rpc_error_response(
-                    None,
-                    JsonRpcError::invalid_request(format!("Invalid request batch: {error}")),
-                );
-            }
-        };
-
-        let mut extensions = crate::router::Extensions::new();
-        extensions.insert(state.protocol_support.clone());
-        extensions.insert(session_revision);
-        #[cfg(feature = "oauth")]
-        if let Some(claims) = http_extensions.get::<crate::oauth::token::TokenClaims>() {
-            extensions.insert(claims.clone());
-        }
-        crate::transport::extension_bridge::apply_extension_bridges(
-            &state.extension_bridges,
-            &http_extensions,
-            &mut extensions,
-        );
-
-        let mut service = JsonRpcService::new(session.make_service())
-            .with_extensions(extensions)
-            .protocol_support(state.protocol_support.clone())
-            .with_negotiated_protocol_version(&session_protocol_version);
-        let response = match service.call_message(message).await {
-            Ok(response) => response,
-            Err(error) => {
-                return json_rpc_error_response(
-                    None,
-                    JsonRpcError::internal_error(error.to_string()),
-                );
-            }
-        };
-        let mut response = axum::Json(response).into_response();
-        response.headers_mut().insert(
-            MCP_PROTOCOL_VERSION_HEADER,
-            HeaderValue::from_str(&session_protocol_version).unwrap(),
-        );
-        return response;
-    }
-
-    // SEP-2575 / SEP-2567: intercept `subscriptions/listen` before the standard
-    // version validation. `subscriptions/listen` is only available when the
-    // effective protocol version is >= 2026-07-28; otherwise we return a
-    // proper JSON-RPC error rather than silently falling through to the
-    // router (which would return `MethodNotFound` anyway, but without the
-    // protocol-version context).
-    //
-    // We check the Mcp-Protocol-Version header first (per-request override),
-    // falling back to the session-negotiated version. Intercepting here
-    // also prevents the version-validation guard below from rejecting the
-    // 2026-07-28 header before we can inspect it.
-    {
-        let method_str = parsed.get("method").and_then(|m| m.as_str()).unwrap_or("");
-        if method_str == "subscriptions/listen" {
-            let req_id = extract_request_id(&parsed);
-            let effective_version = if let Some(v) = get_protocol_version(&headers) {
-                v
-            } else {
-                session.protocol_version.read().await.clone()
-            };
-            if version_supports_subscriptions_listen(&effective_version, &state.protocol_support) {
-                return handle_subscriptions_listen_sse(session).await;
-            } else {
-                return json_rpc_error_response(
-                    req_id,
-                    JsonRpcError::method_not_found("subscriptions/listen"),
-                );
-            }
-        }
-    }
-
-    // SEP-2243: validate the standardized HTTP headers (Mcp-Method,
-    // Mcp-Name, Mcp-Param-*) against the body. Mode is "strict" only
-    // when the negotiated protocol version is at or beyond the
-    // SEP-2243-inclusion version; otherwise present headers are still
-    // checked for body consistency but missing headers are allowed.
-    //
-    // For `initialize` requests the session's protocol version hasn't
-    // been negotiated yet, so we fall back to the version the client
-    // requested in the body. For all other requests we use the session's
-    // negotiated version (which is also reflected back in the response
-    // `Mcp-Protocol-Version` header).
-    let sep_2243_version = if is_init {
-        match parsed
-            .get("params")
-            .and_then(|p| p.get("protocolVersion"))
-            .and_then(|v| v.as_str())
-        {
-            Some(v) => v.to_string(),
-            None => session.protocol_version.read().await.clone(),
-        }
-    } else {
-        session.protocol_version.read().await.clone()
-    };
-    let sep_2243_mode = super::http_headers::mode_for_version(&sep_2243_version);
-    if let Err(err) = super::http_headers::validate_with_tool_schema(
-        &headers,
-        &parsed,
-        sep_2243_mode,
-        tool_input_schema.as_ref(),
-    ) {
-        tracing::warn!(
-            mode = ?sep_2243_mode,
-            version = %sep_2243_version,
-            error = %err.message,
-            "Rejecting request: SEP-2243 header validation failed",
-        );
-        let id = extract_request_id(&parsed);
-        let mut resp = json_rpc_error_response(id, err);
-        // Per SEP-2243 §"Error Code" the HTTP status MUST be 400.
-        *resp.status_mut() = StatusCode::BAD_REQUEST;
-        return resp;
-    }
-
-    // Check if this is a response to one of our outgoing requests (sampling)
-    if is_response(&parsed) {
-        if let Some(id) = extract_request_id(&parsed) {
-            let result = if let Some(error) = parsed.get("error") {
-                let code = error.get("code").and_then(|c| c.as_i64()).unwrap_or(-1);
-                let message = error
-                    .get("message")
-                    .and_then(|m| m.as_str())
-                    .unwrap_or("Unknown error");
-                Err(Error::Internal(format!(
-                    "Client error ({}): {}",
-                    code, message
-                )))
-            } else if let Some(result) = parsed.get("result") {
-                Ok(result.clone())
-            } else {
-                Err(Error::Internal(
-                    "Response has neither result nor error".to_string(),
-                ))
-            };
-
-            if session.complete_pending_request(&id, result).await {
-                tracing::debug!(request_id = ?id, "Completed pending request");
-            } else {
-                tracing::warn!(request_id = ?id, "Received response for unknown request");
-            }
-        }
-        return StatusCode::ACCEPTED.into_response();
-    }
-
-    // Check if this is a notification (no id field)
-    if parsed.get("id").is_none() {
-        // Handle notification
-        if let Ok(notification) = serde_json::from_value::<JsonRpcNotification>(parsed)
-            && let Ok(mcp_notification) = McpNotification::from_jsonrpc(&notification)
-        {
-            // Per the MCP 2025-11-25 spec, clients MUST send
-            // `notifications/initialized` after receiving the `initialize`
-            // response and before sending any other requests. Record the
-            // receipt so the strict_initialization check below can allow
-            // subsequent tool/resource/prompt requests.
-            if matches!(&mcp_notification, McpNotification::Initialized) {
-                session
-                    .initialized_notification_received
-                    .store(true, Ordering::Release);
-                tracing::debug!(session_id = %session.id, "Received notifications/initialized");
-            }
-            session.handle_notification(mcp_notification);
-        }
-        return StatusCode::ACCEPTED.into_response();
-    }
-
-    // Enforce `notifications/initialized` before any non-initialize request
-    // (MCP 2025-11-25 spec requirement). This only applies to the session-based
-    // path; stateless requests (2026-07-28) are handled above and never reach here.
-    if !is_init
-        && state.strict_initialization
-        && !session
-            .initialized_notification_received
-            .load(Ordering::Acquire)
-    {
-        let id = extract_request_id(&parsed);
-        tracing::warn!(
-            session_id = %session.id,
-            "Rejecting request: notifications/initialized not yet received"
-        );
-        return json_rpc_error_response(
-            id,
-            JsonRpcError::invalid_request(
-                "Client must send notifications/initialized before making requests",
-            ),
-        );
-    }
-
-    // For initialize requests, capture the advertised client info /
-    // capabilities from the raw params before `parsed` is consumed by
-    // deserialization. These are stashed onto the live `Session` after a
-    // successful initialize so the persisted SessionRecord faithfully
-    // describes the client (rather than carrying the defaults set at
-    // session-create time).
-    let init_client_metadata: Option<(Option<Implementation>, Option<ClientCapabilities>)> =
-        if is_init {
-            let params = parsed.get("params");
-            let client_info = params
-                .and_then(|p| p.get("clientInfo"))
-                .and_then(|v| serde_json::from_value::<Implementation>(v.clone()).ok());
-            let client_capabilities = params
-                .and_then(|p| p.get("capabilities"))
-                .and_then(|v| serde_json::from_value::<ClientCapabilities>(v.clone()).ok());
-            Some((client_info, client_capabilities))
-        } else {
-            None
-        };
-
-    // Handle as JSON-RPC request
-    let request: JsonRpcRequest = match serde_json::from_value(parsed) {
-        Ok(r) => r,
-        Err(e) => {
-            return json_rpc_error_response(
-                None,
-                JsonRpcError::parse_error(format!("Invalid request: {}", e)),
-            );
-        }
-    };
-
-    // Process the request through the middleware-wrapped service
-    let mut service = JsonRpcService::new(session.make_service());
-
-    // Bridge per-request data from HTTP into MCP Extensions: OAuth claims,
-    // SEP-2575 `_meta` (clientInfo, clientCapabilities, etc.). Empty ext is
-    // skipped to avoid pointless allocation.
-    #[allow(unused_mut)]
-    let mut ext = crate::router::Extensions::new();
-    ext.insert(state.protocol_support.clone());
-    ext.insert(session_revision);
-    #[cfg(feature = "oauth")]
-    if let Some(claims) = http_extensions.get::<crate::oauth::token::TokenClaims>() {
-        ext.insert(claims.clone());
-    }
-    crate::transport::extension_bridge::apply_extension_bridges(
-        &state.extension_bridges,
-        &http_extensions,
-        &mut ext,
-    );
-    #[cfg(feature = "stateless")]
-    stash_per_request_meta(&request, &mut ext);
-
-    // SEP-2260: legacy server-to-client requests are associated with the
-    // client POST that caused them. Give this request its own channel while
-    // drawing IDs from the session-wide allocator so concurrent POSTs cannot
-    // collide or leak requests onto one another's response streams.
-    let mut associated_request_rx = if !is_init {
-        session.request_id_allocator.as_ref().map(|next_id| {
-            let (request_tx, request_rx) = outgoing_request_channel(32);
-            let requester: ClientRequesterHandle = Arc::new(
-                ChannelClientRequester::with_id_allocator(request_tx, next_id.clone()),
-            );
-            ext.insert(requester);
-            request_rx
-        })
-    } else {
-        None
-    };
-
-    if !ext.is_empty() {
-        service = service.with_extensions(ext);
-    }
-
-    let request_id = request.id.clone();
-    let mut call: AssociatedCall = Box::pin(async move { service.call_single(request).await });
-    let mut response = if let Some(mut request_rx) = associated_request_rx.take() {
-        tokio::select! {
-            result = &mut call => match result {
-                Ok(response) => response,
-                Err(error) => {
-                    return json_rpc_error_response(
-                        Some(request_id),
-                        JsonRpcError::internal_error(error.to_string()),
-                    );
-                }
-            },
-            outgoing = request_rx.recv() => {
-                match outgoing {
-                    Some(outgoing) => {
-                        let negotiated_version = session.protocol_version.read().await.clone();
-                        return associated_request_sse_response(
-                            session,
-                            call,
-                            request_rx,
-                            outgoing,
-                            request_id,
-                            request_method,
-                            negotiated_version,
-                        );
-                    }
-                    None => match call.await {
-                        Ok(response) => response,
-                        Err(error) => {
-                            return json_rpc_error_response(
-                                Some(request_id),
-                                JsonRpcError::internal_error(error.to_string()),
-                            );
-                        }
-                    },
-                }
-            }
-        }
-    } else {
-        match call.await {
-            Ok(response) => response,
-            Err(error) => {
-                return json_rpc_error_response(
-                    Some(request_id),
-                    JsonRpcError::internal_error(error.to_string()),
-                );
-            }
-        }
-    };
-
-    // For successful initialize responses, extract and store the negotiated
-    // protocol version, stash the client's advertised identity / capabilities
-    // on the live session, and persist the now-complete record to the session
-    // store so a restore from a peer instance sees the original client info
-    // instead of defaults.
-    if is_init && let JsonRpcResponse::Result(ref result) = response {
-        if let Some(version) = result
-            .result
-            .get("protocolVersion")
-            .and_then(|v| v.as_str())
-        {
-            *session.protocol_version.write().await = version.to_string();
-        }
-        if let Some((client_info, client_capabilities)) = init_client_metadata {
-            *session.client_info.write().await = client_info;
-            *session.client_capabilities.write().await = client_capabilities;
-        }
-        state.sessions.save_record(&session).await;
-    }
-
-    let negotiated_version = session.protocol_version.read().await.clone();
-    let response_version = if request_method == "server/discover"
-        && state.protocol_support.contains(PROTOCOL_VERSION_2026_07_28)
-    {
-        PROTOCOL_VERSION_2026_07_28
-    } else {
-        &negotiated_version
-    };
-    apply_protocol_result_fields(&mut response, &request_method, response_version);
-
-    // Build response with headers
-    let mut resp = if state.sse_responses {
-        sse_json_response(&response)
-    } else {
-        axum::Json(response).into_response()
-    };
-
-    if is_init {
-        resp.headers_mut().insert(
-            MCP_SESSION_ID_HEADER,
-            HeaderValue::from_str(&session.id).unwrap(),
-        );
-    }
-
-    // Always include the negotiated protocol version header
-    resp.headers_mut().insert(
-        MCP_PROTOCOL_VERSION_HEADER,
-        HeaderValue::from_str(&negotiated_version).unwrap(),
-    );
-
-    resp
-}
-
-/// Keep legacy server-to-client requests on the POST response stream that
-/// caused them. These events deliberately have no SSE IDs and are not written
-/// to the session event store: their response channels only exist on this
-/// process and replaying them on another connection would break association.
-fn associated_request_sse_response(
-    session: Arc<Session>,
-    mut call: AssociatedCall,
-    mut request_rx: OutgoingRequestReceiver,
-    first_outgoing: OutgoingRequest,
-    original_request_id: RequestId,
-    request_method: String,
-    negotiated_version: String,
-) -> Response {
-    let (event_tx, event_rx) =
-        tokio::sync::mpsc::channel::<std::result::Result<Event, Infallible>>(32);
-    let call_version = negotiated_version.clone();
-
-    tokio::spawn(async move {
-        let mut pending_ids = Vec::new();
-        if !send_associated_request(&session, &event_tx, first_outgoing, &mut pending_ids).await {
-            session
-                .fail_pending_requests(
-                    &pending_ids,
-                    "originating POST disconnected before the client request was delivered",
-                )
-                .await;
-            return;
-        }
-
-        let mut requests_open = true;
-        loop {
-            tokio::select! {
-                _ = event_tx.closed() => {
-                    session
-                        .fail_pending_requests(
-                            &pending_ids,
-                            "originating POST response stream disconnected",
-                        )
-                        .await;
-                    return;
-                }
-                result = &mut call => {
-                    session
-                        .fail_pending_requests(
-                            &pending_ids,
-                            "originating POST completed before the client request response arrived",
-                        )
-                        .await;
-
-                    let mut response = match result {
-                        Ok(response) => response,
-                        Err(error) => JsonRpcResponse::error(
-                            Some(original_request_id),
-                            JsonRpcError::internal_error(error.to_string()),
-                        ),
-                    };
-                    apply_protocol_result_fields(
-                        &mut response,
-                        &request_method,
-                        &call_version,
-                    );
-
-                    match serde_json::to_string(&response) {
-                        Ok(data) => {
-                            let _ = event_tx
-                                .send(Ok(
-                                    Event::default()
-                                        .event(SSE_MESSAGE_EVENT)
-                                        .data(data),
-                                ))
-                                .await;
-                        }
-                        Err(error) => {
-                            tracing::error!(
-                                error = %error,
-                                "Failed to serialize associated POST response",
-                            );
-                        }
-                    }
-                    return;
-                }
-                outgoing = request_rx.recv(), if requests_open => {
-                    match outgoing {
-                        Some(outgoing) => {
-                            if !send_associated_request(
-                                &session,
-                                &event_tx,
-                                outgoing,
-                                &mut pending_ids,
-                            )
-                            .await
-                            {
-                                session
-                                    .fail_pending_requests(
-                                        &pending_ids,
-                                        "originating POST disconnected before the client request was delivered",
-                                    )
-                                    .await;
-                                return;
-                            }
-                        }
-                        None => requests_open = false,
-                    }
-                }
-            }
-        }
-    });
-
-    let stream = tokio_stream::wrappers::ReceiverStream::new(event_rx);
-    let mut response = Sse::new(stream)
-        .keep_alive(
-            axum::response::sse::KeepAlive::new()
-                .interval(Duration::from_secs(30))
-                .text("ping"),
-        )
-        .into_response();
-    response.headers_mut().insert(
-        MCP_PROTOCOL_VERSION_HEADER,
-        HeaderValue::from_str(&negotiated_version).unwrap(),
-    );
-    response
-}
-
-async fn send_associated_request(
-    session: &Session,
-    event_tx: &tokio::sync::mpsc::Sender<std::result::Result<Event, Infallible>>,
-    outgoing: OutgoingRequest,
-    pending_ids: &mut Vec<RequestId>,
-) -> bool {
-    let id = outgoing.id.clone();
-    let request = JsonRpcRequest {
-        jsonrpc: "2.0".to_string(),
-        id: id.clone(),
-        method: outgoing.method,
-        params: Some(outgoing.params),
-    };
-    let data = match serde_json::to_string(&request) {
-        Ok(data) => data,
-        Err(error) => {
-            let _ = outgoing.response_tx.send(Err(Error::Internal(format!(
-                "Failed to serialize associated client request: {error}"
-            ))));
-            return true;
-        }
-    };
-
-    session
-        .add_pending_request(id.clone(), outgoing.response_tx)
-        .await;
-    pending_ids.push(id);
-
-    event_tx
-        .send(Ok(Event::default().event(SSE_MESSAGE_EVENT).data(data)))
-        .await
-        .is_ok()
-}
-
-/// Returns `true` when the given protocol version string enables `subscriptions/listen`.
-///
-/// `subscriptions/listen` is part of the 2026-07-28 spec (SEP-2575 / SEP-2567).
-/// Unknown future dates do not opt into behavior that has not been compiled
-/// and explicitly enabled.
-fn version_supports_subscriptions_listen(
-    version: &str,
-    protocol_support: &ProtocolSupport,
-) -> bool {
-    version == PROTOCOL_VERSION_2026_07_28 && protocol_support.contains(version)
-}
-
-/// Returns `true` when the given protocol version string enables stateless
-/// (sessionless) mode for the HTTP transport.
-///
-/// Stateless mode is introduced in the 2026-07-28 protocol (SEP-2575 /
-/// SEP-2567). Only the exact, compiled-and-enabled version opts in; unknown
-/// future dates must not silently inherit revision-specific behavior.
+mod stateless_dispatch;
+
+use handlers::{handle_delete, handle_get, handle_health, handle_post};
+pub use session::{DEFAULT_SESSION_TTL, SessionConfig, SessionHandle, SessionInfo};
+// Only `stateless_dispatch` (gated below) calls back into these two; importing
+// them unconditionally would be an unused import (-> error under -D warnings)
+// in a build without the feature.
 #[cfg(feature = "stateless")]
-fn is_stateless_protocol_version(version: &str) -> bool {
-    version == PROTOCOL_VERSION_2026_07_28
-}
-
-/// Stamp `_meta["io.modelcontextprotocol/serverInfo"]` onto a successful
-/// response, per SEP-2575: servers SHOULD identify themselves in each
-/// result's `_meta` unless configured not to (see
-/// [`HttpTransport::stamp_server_info()`]).
-///
-/// A no-op for error responses, and for any result whose top-level JSON
-/// value isn't an object (defensive; every `McpResponse` variant serializes
-/// to an object).
+use handlers::{extract_request_id, json_rpc_error_response_with_status};
+use session::{Session, SessionRegistry};
+// Gated in `stateless_dispatch` too, so importing any of these unconditionally
+// breaks the default build that `--all-features` never exercises.
 #[cfg(feature = "stateless")]
-fn stamp_server_info(response: &mut JsonRpcResponse, implementation: &Implementation) {
-    let JsonRpcResponse::Result(result) = response else {
-        return;
-    };
-    let Some(obj) = result.result.as_object_mut() else {
-        return;
-    };
-    let meta = obj
-        .entry("_meta")
-        .or_insert_with(|| serde_json::Value::Object(Default::default()));
-    let Some(meta_obj) = meta.as_object_mut() else {
-        return;
-    };
-    if let Ok(value) = serde_json::to_value(implementation) {
-        meta_obj.insert("io.modelcontextprotocol/serverInfo".to_string(), value);
-    }
-}
-
-/// Drop guard that cancels a per-request [`CancellationToken`] when the
-/// request is abandoned before its response is produced.
-///
-/// On the sessionless POST path the response future (plain JSON) or the SSE
-/// response stream is dropped when the client disconnects; holding this
-/// guard in that future/stream turns the drop into a cancellation signal.
-/// [`disarm`](Self::disarm) once the handler's terminal response resolves
-/// so normal completion doesn't signal cancellation.
-#[cfg(feature = "stateless")]
-struct CancelOnDisconnect(Option<crate::context::CancellationToken>);
-
-#[cfg(feature = "stateless")]
-impl CancelOnDisconnect {
-    fn arm(token: crate::context::CancellationToken) -> Self {
-        Self(Some(token))
-    }
-
-    fn disarm(&mut self) {
-        self.0 = None;
-    }
-}
-
-#[cfg(feature = "stateless")]
-impl Drop for CancelOnDisconnect {
-    fn drop(&mut self) {
-        if let Some(token) = self.0.take() {
-            token.cancel();
-        }
-    }
-}
-
-/// Stream a sessionless POST response as SSE: the notifications the handler
-/// emitted, in order, followed by the terminal JSON-RPC response.
-///
-/// Invoked when a handler produced a notification before its terminal
-/// response on the 2026-07-28 sessionless path. A plain JSON body would drop
-/// those notifications (there is no session stream to carry them), so the
-/// response falls back to `text/event-stream`: the buffered first
-/// notification, any further notifications as they arrive, and finally the
-/// terminal response, after which the stream ends.
-#[cfg(feature = "stateless")]
-struct StatelessSseContext {
-    version: String,
-    method: String,
-    cancel_guard: CancelOnDisconnect,
-    server_identity: Option<Implementation>,
-    subscriptions: Arc<ModernSubscriptionRegistry>,
-}
-
-#[cfg(feature = "stateless")]
-fn stateless_sse_with_notifications(
-    first: crate::context::ServerNotification,
-    call: std::pin::Pin<
-        Box<dyn std::future::Future<Output = crate::error::Result<JsonRpcResponse>> + Send>,
-    >,
-    rx: crate::context::NotificationReceiver,
-    request: StatelessSseContext,
-) -> Response {
-    struct Ctx {
-        call: Option<
-            std::pin::Pin<
-                Box<dyn std::future::Future<Output = crate::error::Result<JsonRpcResponse>> + Send>,
-            >,
-        >,
-        rx: crate::context::NotificationReceiver,
-        rx_open: bool,
-        queue: std::collections::VecDeque<String>,
-        terminal: Option<String>,
-        version: String,
-        method: String,
-        /// Cancels the per-request token if the client disconnects (the
-        /// stream, and with it this state, is dropped) while the handler
-        /// is still in flight. Disarmed once the handler resolves.
-        cancel_guard: CancelOnDisconnect,
-        /// Stamped into `_meta.serverInfo` on the terminal response, if set
-        /// (see [`HttpTransport::stamp_server_info()`]).
-        server_identity: Option<Implementation>,
-        subscriptions: Arc<ModernSubscriptionRegistry>,
-    }
-
-    let mut queue = std::collections::VecDeque::new();
-    if !request.subscriptions.publish(&first)
-        && let Some(json) = crate::transport::stdio::serialize_notification(&first)
-    {
-        queue.push_back(json);
-    }
-    let ctx = Ctx {
-        call: Some(call),
-        rx,
-        rx_open: true,
-        queue,
-        terminal: None,
-        version: request.version,
-        method: request.method,
-        cancel_guard: request.cancel_guard,
-        server_identity: request.server_identity,
-        subscriptions: request.subscriptions,
-    };
-
-    let stream = futures::stream::unfold(ctx, |mut ctx| async move {
-        loop {
-            // Buffered notifications flush first to preserve emission order.
-            if let Some(json) = ctx.queue.pop_front() {
-                return Some((
-                    Ok::<_, Infallible>(Event::default().event(SSE_MESSAGE_EVENT).data(json)),
-                    ctx,
-                ));
-            }
-            // The terminal response is the last event on the stream.
-            if let Some(json) = ctx.terminal.take() {
-                return Some((
-                    Ok(Event::default().event(SSE_MESSAGE_EVENT).data(json)),
-                    ctx,
-                ));
-            }
-            let mut call = ctx.call.take()?;
-            tokio::select! {
-                result = &mut call => {
-                    // Handler finished; a later disconnect is no longer a
-                    // cancellation.
-                    ctx.cancel_guard.disarm();
-                    // Drain notifications that were queued before the handler
-                    // finished so they precede the terminal response.
-                    while let Ok(n) = ctx.rx.try_recv() {
-                        if !ctx.subscriptions.publish(&n)
-                            && let Some(json) =
-                                crate::transport::stdio::serialize_notification(&n)
-                        {
-                            ctx.queue.push_back(json);
-                        }
-                    }
-                    let terminal_json = match result {
-                        Ok(mut response) => {
-                            // Same initialize version patch as the JSON path.
-                            if ctx.method == "initialize"
-                                && let JsonRpcResponse::Result(ref mut r) = response
-                                && let Some(pv) = r.result.get_mut("protocolVersion")
-                            {
-                                *pv = serde_json::Value::String(ctx.version.clone());
-                            }
-                            apply_protocol_result_fields(
-                                &mut response,
-                                &ctx.method,
-                                &ctx.version,
-                            );
-                            if let Some(ref identity) = ctx.server_identity {
-                                stamp_server_info(&mut response, identity);
-                            }
-                            serde_json::to_string(&response).ok()
-                        }
-                        Err(e) => Some(
-                            serde_json::json!({
-                                "jsonrpc": "2.0",
-                                "id": serde_json::Value::Null,
-                                "error": JsonRpcError::internal_error(e.to_string()),
-                            })
-                            .to_string(),
-                        ),
-                    };
-                    ctx.terminal = terminal_json;
-                    // `call` is complete and intentionally not restored.
-                }
-                maybe = ctx.rx.recv(), if ctx.rx_open => {
-                    match maybe {
-                        Some(n) => {
-                            if !ctx.subscriptions.publish(&n)
-                                && let Some(json) =
-                                    crate::transport::stdio::serialize_notification(&n)
-                            {
-                                ctx.queue.push_back(json);
-                            }
-                        }
-                        None => ctx.rx_open = false,
-                    }
-                    ctx.call = Some(call);
-                }
-            }
-        }
-    });
-
-    Sse::new(stream)
-        .keep_alive(
-            axum::response::sse::KeepAlive::new()
-                .interval(Duration::from_secs(30))
-                .text("ping"),
-        )
-        .into_response()
-}
-
-/// Serve the final, sessionless `subscriptions/listen` protocol over its
-/// owning POST response.
-#[cfg(feature = "stateless")]
-async fn handle_modern_subscriptions_listen_sse(
-    state: Arc<AppState>,
-    parsed: &serde_json::Value,
-    http_extensions: &axum::http::Extensions,
-) -> Response {
-    let id = extract_request_id(parsed);
-    let Some(subscription_id) = id.clone() else {
-        return json_rpc_error_response_with_status(
-            None,
-            JsonRpcError::invalid_request("subscriptions/listen requires a request id"),
-            StatusCode::BAD_REQUEST,
-        );
-    };
-    let request: JsonRpcRequest = match serde_json::from_value(parsed.clone()) {
-        Ok(request) => request,
-        Err(error) => {
-            return json_rpc_error_response_with_status(
-                id,
-                JsonRpcError::invalid_request(format!("Invalid request: {error}")),
-                StatusCode::BAD_REQUEST,
-            );
-        }
-    };
-
-    // Dispatch the request through the per-request service before upgrading,
-    // so `Service<RouterRequest>` middleware observes accepted and rejected
-    // listens and the router owns validation and filter negotiation (#1182).
-    // The service response never reaches the wire: an error is returned as
-    // the reply, and a success carries the accepted filter this handler
-    // consumes to register the stream. The stream lifetime stays entirely
-    // transport-owned.
-    let service = match &state.service_source {
-        ServiceSource::Router { router, factory } => {
-            let ephemeral = router.with_fresh_session();
-            ephemeral.session().mark_preinitialized();
-            JsonRpcService::new(factory(ephemeral))
-        }
-        ServiceSource::Service(mutex) => JsonRpcService::new(mutex.lock().unwrap().clone()),
-    };
-    let mut ext = crate::router::Extensions::new();
-    ext.insert(state.protocol_support.clone());
-    stash_per_request_meta(&request, &mut ext);
-    crate::transport::extension_bridge::apply_extension_bridges(
-        &state.extension_bridges,
-        http_extensions,
-        &mut ext,
-    );
-    if ext
-        .get::<crate::stateless::StatelessRequestMeta>()
-        .is_none()
-    {
-        // This handler is only reached for effective-final requests, but the
-        // version can arrive via the HTTP header with no per-request `_meta`.
-        // Seed the meta so the router classifies the request correctly.
-        ext.insert(crate::stateless::StatelessRequestMeta {
-            protocol_version: Some(crate::protocol::PROTOCOL_VERSION_2026_07_28.to_string()),
-            ..Default::default()
-        });
-    }
-    let mut service = service.with_extensions(ext);
-
-    let response = match service.call_single(request).await {
-        Ok(response) => response,
-        Err(error) => {
-            return json_rpc_error_response_with_status(
-                id,
-                JsonRpcError::internal_error(error.to_string()),
-                StatusCode::INTERNAL_SERVER_ERROR,
-            );
-        }
-    };
-    let accepted = match &response {
-        JsonRpcResponse::Result(result) => match result
-            .result
-            .get("notifications")
-            .cloned()
-            .map(serde_json::from_value::<SubscriptionFilter>)
-        {
-            Some(Ok(accepted)) => accepted,
-            _ => {
-                return json_rpc_error_response_with_status(
-                    id,
-                    JsonRpcError::internal_error(
-                        "subscriptions/listen produced an unrecognized service result",
-                    ),
-                    StatusCode::INTERNAL_SERVER_ERROR,
-                );
-            }
-        },
-        _ => {
-            // Rejected: middleware already observed the error; reply with it.
-            let mut resp = axum::Json(&response).into_response();
-            *resp.status_mut() = StatusCode::BAD_REQUEST;
-            return resp;
-        }
-    };
-
-    let (rx, guard) = state
-        .modern_subscriptions
-        .register(subscription_id.clone(), accepted.clone());
-    let acknowledgment = serde_json::json!({
-        "jsonrpc": "2.0",
-        "method": "notifications/subscriptions/acknowledged",
-        "params": {
-            "_meta": {
-                "io.modelcontextprotocol/subscriptionId": subscription_id
-            },
-            "notifications": accepted
-        }
-    })
-    .to_string();
-
-    struct ModernListenStream {
-        first: Option<String>,
-        rx: mpsc::UnboundedReceiver<String>,
-        _guard: ModernSubscriptionGuard,
-    }
-
-    let stream = futures::stream::unfold(
-        ModernListenStream {
-            first: Some(acknowledgment),
-            rx,
-            _guard: guard,
-        },
-        |mut state| async move {
-            let message = match state.first.take() {
-                Some(first) => Some(first),
-                None => state.rx.recv().await,
-            }?;
-            Some((
-                Ok::<_, Infallible>(Event::default().event(SSE_MESSAGE_EVENT).data(message)),
-                state,
-            ))
-        },
-    );
-
-    let mut response = Sse::new(stream)
-        .keep_alive(
-            axum::response::sse::KeepAlive::new()
-                .interval(Duration::from_secs(30))
-                .text("ping"),
-        )
-        .into_response();
-    response.headers_mut().insert(
-        MCP_PROTOCOL_VERSION_HEADER,
-        HeaderValue::from_static(PROTOCOL_VERSION_2026_07_28),
-    );
-    response
-}
-
-/// Serve a `subscriptions/listen` request as an SSE stream.
-///
-/// Subscribes to the session's notification broadcast channel and returns a
-/// streaming `text/event-stream` response. The stream closes naturally when:
-/// - The client disconnects (axum drops the response body).
-/// - The broadcast channel closes (server shutdown / session expiry).
-///
-/// Each notification is assigned a monotonically increasing event ID for
-/// potential stream resumption (SEP-1699).
-async fn handle_subscriptions_listen_sse(session: Arc<Session>) -> Response {
-    let rx = session.notifications_tx.subscribe();
-    let session_clone = session.clone();
-
-    let stream = BroadcastStream::new(rx)
-        .then(move |result: std::result::Result<String, _>| {
-            let session = session_clone.clone();
-            async move {
-                match result {
-                    Ok(msg) => {
-                        let event_id = session.next_event_id();
-                        // Buffer the event for potential replay (SEP-1699)
-                        session.buffer_event(event_id, msg.clone()).await;
-                        Some(Ok::<_, Infallible>(
-                            Event::default()
-                                .id(event_id.to_string())
-                                .event(SSE_MESSAGE_EVENT)
-                                .data(msg),
-                        ))
-                    }
-                    Err(_) => None,
-                }
-            }
-        })
-        .filter_map(|x| x);
-
-    Sse::new(stream)
-        .keep_alive(
-            axum::response::sse::KeepAlive::new()
-                .interval(Duration::from_secs(30))
-                .text("ping"),
-        )
-        .into_response()
-}
-
-/// Handle GET requests (SSE stream for server notifications and outgoing requests)
-async fn handle_get(
-    State(state): State<Arc<AppState>>,
-    request: axum::extract::Request,
-) -> Response {
-    let (parts, _body) = request.into_parts();
-    let headers = parts.headers;
-    let uri = parts.uri.clone();
-
-    // Validate Host (DNS rebinding defense, complement to Origin)
-    if let Some(resp) = validate_host(&headers, &uri, &state) {
-        return resp;
-    }
-
-    // Validate Origin
-    if let Some(resp) = validate_origin(&headers, &state) {
-        return resp;
-    }
-
-    // Check Accept header
-    let accept = headers
-        .get(header::ACCEPT)
-        .and_then(|v| v.to_str().ok())
-        .unwrap_or("");
-
-    if !accept.contains("text/event-stream") {
-        return (
-            StatusCode::NOT_ACCEPTABLE,
-            "Accept header must include text/event-stream",
-        )
-            .into_response();
-    }
-
-    // Get session
-    let session_id = match get_session_id(&headers) {
-        Some(id) => id,
-        None => {
-            return json_rpc_error_response(None, JsonRpcError::session_required());
-        }
-    };
-
-    let session = match state.sessions.get(&session_id).await {
-        Some(s) => s,
-        None => {
-            return json_rpc_error_response(
-                None,
-                JsonRpcError::session_not_found_with_id(&session_id),
-            );
-        }
-    };
-
-    // Check for Last-Event-ID header for stream resumption (SEP-1699)
-    let last_event_id = get_last_event_id(&headers);
-
-    // GET is the resumable notification stream. Restricted server-to-client
-    // requests are emitted only on their originating POST response stream.
-    let rx = session.notifications_tx.subscribe();
-    let session_clone = session.clone();
-
-    // Replay buffered events if Last-Event-ID was provided (SEP-1699)
-    let replay_events: Vec<_> = if let Some(after_id) = last_event_id {
-        let events = session.get_events_after(after_id).await;
-        tracing::debug!(
-            after_id = after_id,
-            replay_count = events.len(),
-            "Replaying buffered events for stream resumption"
-        );
-        events
-            .into_iter()
-            .map(|e| {
-                Ok::<_, Infallible>(
-                    Event::default()
-                        .id(e.id.to_string())
-                        .event(SSE_MESSAGE_EVENT)
-                        .data(e.data),
-                )
-            })
-            .collect()
-    } else {
-        Vec::new()
-    };
-
-    // Create replay stream from buffered events
-    let replay_stream = tokio_stream::iter(replay_events);
-
-    // Create live stream for new events
-    // Use `then` for async processing, then `filter_map` to remove errors
-    let live_stream = BroadcastStream::new(rx)
-        .then(move |result: std::result::Result<String, _>| {
-            let session = session_clone.clone();
-            async move {
-                match result {
-                    Ok(msg) => {
-                        let event_id = session.next_event_id();
-                        // Buffer the event for potential replay (SEP-1699)
-                        session.buffer_event(event_id, msg.clone()).await;
-                        Some(Ok::<_, Infallible>(
-                            Event::default()
-                                .id(event_id.to_string())
-                                .event(SSE_MESSAGE_EVENT)
-                                .data(msg),
-                        ))
-                    }
-                    Err(_) => None,
-                }
-            }
-        })
-        .filter_map(|x| x);
-
-    // Chain replay stream with live stream
-    let stream = replay_stream.chain(live_stream);
-
-    Sse::new(stream)
-        .keep_alive(
-            axum::response::sse::KeepAlive::new()
-                .interval(Duration::from_secs(30))
-                .text("ping"),
-        )
-        .into_response()
-}
-
-/// Handle DELETE requests (session termination)
-async fn handle_delete(
-    State(state): State<Arc<AppState>>,
-    request: axum::extract::Request,
-) -> Response {
-    let (parts, _body) = request.into_parts();
-    let headers = parts.headers;
-    let uri = parts.uri.clone();
-
-    // Validate Host (DNS rebinding defense, complement to Origin)
-    if let Some(resp) = validate_host(&headers, &uri, &state) {
-        return resp;
-    }
-
-    // Validate Origin
-    if let Some(resp) = validate_origin(&headers, &state) {
-        return resp;
-    }
-
-    let session_id = match get_session_id(&headers) {
-        Some(id) => id,
-        None => {
-            return json_rpc_error_response(None, JsonRpcError::session_required());
-        }
-    };
-
-    if state.sessions.remove(&session_id).await {
-        tracing::info!(session_id = %session_id, "Session terminated");
-        StatusCode::OK.into_response()
-    } else {
-        // For DELETE, it's okay if the session doesn't exist - it's already gone
-        // Return OK instead of an error for idempotency
-        tracing::debug!(session_id = %session_id, "Session already removed or never existed");
-        StatusCode::OK.into_response()
-    }
-}
-
-/// Handle GET /health requests
-///
-/// Returns a simple 200 OK response for health checks.
-/// Does not require authentication or session state.
-async fn handle_health() -> Response {
-    StatusCode::OK.into_response()
-}
-
-/// Build a synchronous JSON-RPC response wrapped in SSE format.
-///
-/// Used when [`AppState::sse_responses`] is `true`. The body is a single SSE
-/// event followed by the required blank line:
-///
-/// ```text
-/// event: message
-/// data: <json>
-///
-/// ```
-fn sse_json_response(response: impl serde::Serialize) -> Response {
-    let json = match serde_json::to_string(&response) {
-        Ok(s) => s,
-        Err(e) => {
-            tracing::error!(error = %e, "Failed to serialize response for SSE wrapping");
-            return StatusCode::INTERNAL_SERVER_ERROR.into_response();
-        }
-    };
-    let sse_body = format!("event: message\ndata: {json}\n\n");
-    (
-        StatusCode::OK,
-        [
-            (header::CONTENT_TYPE, "text/event-stream"),
-            (header::CACHE_CONTROL, "no-cache"),
-        ],
-        sse_body,
-    )
-        .into_response()
-}
-
-/// Create a JSON-RPC error response
-fn json_rpc_error_response(
-    id: Option<crate::protocol::RequestId>,
-    error: JsonRpcError,
-) -> Response {
-    let response = JsonRpcResponse::error(id, error);
-    axum::Json(response).into_response()
-}
-
-fn json_rpc_error_response_with_status(
-    id: Option<crate::protocol::RequestId>,
-    error: JsonRpcError,
-    status: StatusCode,
-) -> Response {
-    let mut response = json_rpc_error_response(id, error);
-    *response.status_mut() = status;
-    response
-}
-
-/// HTTP 413 response for a POST body exceeding [`HttpTransport::max_body_size`].
-fn body_too_large_response(limit: usize) -> Response {
-    let mut resp = json_rpc_error_response(
-        None,
-        JsonRpcError::invalid_request(format!(
-            "Request body exceeds the maximum size of {} bytes",
-            limit
-        )),
-    );
-    *resp.status_mut() = StatusCode::PAYLOAD_TOO_LARGE;
-    resp
-}
-
-/// Returns `true` when the body-read error was caused by exceeding the
-/// configured length limit (as opposed to a transport-level I/O failure).
-fn is_length_limit_error(err: &axum::Error) -> bool {
-    let mut source: Option<&(dyn std::error::Error + 'static)> = Some(err);
-    while let Some(e) = source {
-        if e.is::<http_body_util::LengthLimitError>() {
-            return true;
-        }
-        source = e.source();
-    }
-    false
-}
+use stateless_dispatch::{
+    CancelOnDisconnect, ModernSubscriptionRegistry, StatelessSseContext,
+    handle_modern_subscriptions_listen_sse, is_stateless_protocol_version, modern_response_status,
+    stamp_server_info, stash_per_request_meta, stateless_sse_with_notifications,
+};
 
 #[cfg(test)]
 mod tests;
